% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefactorPriorKnoweldge.R
\name{AddInfo}
\alias{AddInfo}
\title{Adds extra columns to enrichment output that inform about 1. The amount of genes associated with term in prior knowledge, 2. The amount of genes detected in input data associated with term in prior knowledge, and 3. The percentage of genes detected in input data associated with term in prior knowledge.}
\usage{
AddInfo(mat, net, res, .source, .target, complete = FALSE)
}
\arguments{
\item{mat}{Data matrix used as input for enrichment analysis}

\item{net}{Prior Knowledge used as input for enrichment analysis}

\item{res}{Results returned from the enrichment analysis}

\item{.source}{used as input for enrichment analysis}

\item{.target}{used as input for enrichment analysis}

\item{complete}{TRUE or FALSE, weather only .source with results should be returned or all .source in net.}
}
\description{
Adds extra columns to enrichment output that inform about 1. The amount of genes associated with term in prior knowledge, 2. The amount of genes detected in input data associated with term in prior knowledge, and 3. The percentage of genes detected in input data associated with term in prior knowledge.
}
