% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyData.R
\docType{data}
\name{alanine_pathways}
\alias{alanine_pathways}
\title{alanine_pathways}
\format{
Columns include pathway_name, pathwayId and pathwaySource as well as inputID  and commonName.
}
\source{
Schmidt et al, MetaProViz: METabolomics pre-PRocessing, functiOnal analysis and VIZualisation version 2.1.7, GitHub 2025.
}
\usage{
alanine_pathways
}
\description{
Manually curated table for the amino acid alanine toshowcase pathways (wiki, reactome, etc.) and alanine IDs (chebi, hmdb, etc.)
included in those pathways
}
\examples{
data(alanine_pathways)
head(alanine_pathways)
}
\keyword{datasets}
