% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyData.R
\docType{data}
\name{biocrates_features}
\alias{biocrates_features}
\title{biocrates_features}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1019 rows and 14 columns.
}
\usage{
biocrates_features
}
\description{
Biocrates kit feature information of the "MxP Quant 500 XL kit" that covers
more than 1,000 metabolites with biochemical class information and the
exported different metabolite IDs (HMDB, KEGG, etc.). information (INCHI,
Key, etc.).
}
\examples{
data(biocrates_features)
head(biocrates_features)

}
\keyword{datasets}
