% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverRepresentationAnalysis.R
\name{cluster_ora}
\alias{cluster_ora}
\title{\subsection{---------------------------}{

Overrepresentation analysis by cluster
}}
\usage{
cluster_ora(
  data,
  metadata_info = c(ClusterColumn = "RG2_Significant", BackgroundColumn = "BG_method",
    PathwayTerm = "term", PathwayFeature = "Metabolite"),
  remove_background = TRUE,
  input_pathway,
  pathway_name = "",
  min_gssize = 10,
  max_gssize = 1000,
  save_table = "csv",
  path = NULL
)
}
\arguments{
\item{data}{DF with metabolite names/metabolite IDs as row names. Metabolite names/IDs need to match the identifier type (e.g. HMDB IDs) in the input_pathway.}

\item{metadata_info}{\emph{Optional: } Pass ColumnName of the column including the cluster names that ORA should be performed on (=ClusterColumn). BackgroundColumn passes the column name needed if remove_background=TRUE. Also pass ColumnName for input_pathway including term and feature names. (ClusterColumn= ColumnName data, BackgroundColumn = ColumnName data, PathwayTerm= ColumnName input_pathway, PathwayFeature= ColumnName input_pathway) \strong{c(FeatureName="Metabolite", ClusterColumn="RG2_Significant", BackgroundColumn="BG_method", PathwayTerm= "term", PathwayFeature= "Metabolite")}}

\item{remove_background}{\emph{Optional: } If TRUE, column BackgroundColumn  name needs to be in metadata_info, which includes TRUE/FALSE for each metabolite to fall into background based on the chosen Background method for e.g. mca_2cond are removed from the universe. \strong{default: TRUE}}

\item{input_pathway}{DF that must include column "term" with the pathway name, column "Feature" with the Metabolite name or ID and column "Description" with pathway description.}

\item{pathway_name}{\emph{Optional: } Name of the pathway list used \strong{default: ""}}

\item{min_gssize}{\emph{Optional: } minimum group size in ORA \strong{default: 10}}

\item{max_gssize}{\emph{Optional: } maximum group size in ORA \strong{default: 1000}}

\item{save_table}{\emph{Optional: } File types for the analysis results are: "csv", "xlsx", "txt" \strong{default: "csv"}}

\item{path}{\emph{Optional:} Path to the folder the results should be saved at. \strong{default: NULL}}
}
\value{
Saves results as individual .csv files.
}
\description{
Uses enricher to run ORA on each of the metabolite cluster from any of the MCA functions using a pathway list
}
\examples{
KEGG_Pathways <- MetaProViz::metsigdb_kegg()
DMAres <- as.data.frame(intracell_dma) \%>\% dplyr::filter(!is.na(KEGGCompound))\%>\% as.data.frame()\%>\%tibble::column_to_rownames("KEGGCompound")\%>\%dplyr::select(- "Metabolite")
MetaProViz::cluster_ora(data= DMAres,
metadata_info=c(ClusterColumn="Pathway", PathwayTerm= "term", PathwayFeature= "Metabolite"),
input_pathway=KEGG_Pathways,
remove_background=FALSE)

}
