% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefactorPriorKnoweldge.R
\name{ClusterPK}
\alias{ClusterPK}
\title{Deal with detected input features. If we only have one ID, could it also be another one?}
\usage{
ClusterPK(
  InputData,
  SettingsInfo = c(InputID = "MetaboliteID", GroupingVariable = "term"),
  Clust = "Graph",
  matrix = "percentage",
  min = 2
)
}
\arguments{
\item{InputData}{Dataframe with at least one column with the target (e.g. metabolite) and a column source (e.g. term).}

\item{SettingsInfo}{= c(InputID="MetaboliteID", GroupingVariable="term"),}
}
\description{
Deal with pathway overlap in prior knowledge
}
\examples{
KEGG_Pathways <- MetaProViz::LoadKEGG()
InputData = KEGG_Pathways



}
