% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefactorPriorKnoweldge.R
\name{count_id}
\alias{count_id}
\title{Count Entries and Generate a Histogram Plot for a Specified Column}
\usage{
count_id(
  data,
  column,
  delimiter = ",",
  fill_colors = c(`No ID` = "#FB8072", `Single ID` = "#B3DE69", `Multiple IDs` =
    "#80B1D3"),
  binwidth = 1,
  title_prefix = NULL,
  save_plot = "svg",
  save_table = "csv",
  print_plot = TRUE,
  path = NULL
)
}
\arguments{
\item{data}{A data frame containing the data to be analyzed.}

\item{column}{A string specifying the name of the column in \code{data} to
analyze.}

\item{delimiter}{A string specifying the delimiter used to split cell
values.
Defaults to \code{","}.}

\item{fill_colors}{A named character vector providing colors for each
category. Defaults to
\code{c("No ID" = "#FB8072", "Single ID" = "#B3DE69", "Multiple IDs" =
"#80B1D3")}.}

\item{binwidth}{Numeric value specifying the bin width for the histogram.
Defaults to \code{1}.}

\item{title_prefix}{A string to use as the title of the plot. If \code{NULL}
(default), the title
will be generated as "Number of \if{html}{\out{<column>}} IDs per Biocrates Cell".}

\item{save_plot}{\emph{Optional: } Select the file type of output plots.
Options are svg, png, pdf. \strong{Default = svg}}

\item{save_table}{\emph{Optional: } File types for the analysis results are:
"csv", "xlsx", "txt". \strong{Default = "csv"}}

\item{print_plot}{\emph{Optional: } TRUE or FALSE, if TRUE Volcano plot is
saved as an overview of the results. \strong{Default = TRUE}}

\item{path}{\emph{Optional:} Path to the folder the results should be saved
at. \strong{Default = NULL}}
}
\value{
A list with two elements:
\item{result}{A data frame that includes three additional columns:
\code{was_na} (logical indicator
of missing or empty cells), \code{entry_count} (number of entries in each
cell), and
\code{id_label} (a categorical label based on the entry count).}
\item{plot}{A \code{ggplot} object representing the histogram of entry
counts.}
}
\description{
This function processes a data frame column by counting the number of
entries
within each cell.
It considers both \code{NA} values and empty strings as zero entries, and
categorizes each cell as
"No ID", "Single ID", or "Multiple IDs" based on the count. A histogram is
then generated to visualize
the distribution of entry counts.
}
\examples{
data(biocrates_features)
count_id(biocrates_features, "HMDB")

}
