% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefactorPriorKnoweldge.R
\name{DetectedID}
\alias{DetectedID}
\title{Clean translated ID in prior knowledge based on measured features}
\usage{
DetectedID(
  InputData,
  SettingsInfo = c(InputID = "MetaboliteID", GroupingVariable = "term")
)
}
\arguments{
\item{InputData}{Dataframe with at least one column with the target (e.g. metabolite), you can add other columns such as source (e.g. term)}

\item{SettingsInfo}{}
}
\description{
Clean translated ID in prior knowledge based on measured features
}
