% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverRepresentationAnalysis.R
\name{DM_ORA}
\alias{DM_ORA}
\title{DM_ORA}
\usage{
DM_ORA(
  Input_data,
  PathwayFile,
  PathwayName = "",
  minGSSize = 10,
  maxGSSize = 1000,
  pCutoff = 0.05,
  PercentageCutoff = 10,
  Save_as_Results = "csv",
  Folder_Name = NULL
)
}
\arguments{
\item{Input_data}{Input DF with column "t.val" and column "Metabolite", which needs to match the identifier type in column "Metabolite" in the PathwayFile.}

\item{PathwayFile}{DF that must include column "term" with the pathway name, column "Metabolite" with the Metabolite name or ID and column "Description" with pathway description that will be depicted on the plots.}

\item{PathwayName}{\emph{Optional: } Name of the pathway list used \strong{default: ""}}

\item{minGSSize}{\emph{Optional: } minimum group size in ORA \strong{default: 10}}

\item{maxGSSize}{\emph{Optional: } maximum group size in ORA \strong{default: 1000}}

\item{pCutoff}{\emph{Optional: } p-adjusted value cutoff from ORA results. \strong{default: 0.05}}

\item{PercentageCutoff}{\emph{Optional: } Percentage cutoff of metabolites that should be considered for ORA. \strong{default: 10}}

\item{Save_as_Results}{\emph{Optional: } File types for the analysis results are: "csv", "xlsx", "txt" \strong{default: "csv"}}

\item{Folder_Name}{{Optional:} String which is added to the resulting folder name \strong(Default = NULL)}
}
\value{
Saves results as individual .csv files.
}
\description{
Uses enricher to run ORA on the differential metabolites (DM) using a pathway list
}
