% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyData.R
\docType{data}
\name{equivalent_features}
\alias{equivalent_features}
\title{equivalent_features}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 34 rows and 7 columns.
}
\usage{
equivalent_features
}
\description{
Manually curated list of aminoacids and aminoacid-related metabolites with
corresponding metabolite identifiers (HMDB, KEGG, etc.) irrespective of
chirality.
}
\examples{
data(equivalent_features)
head(equivalent_features)

}
\keyword{datasets}
