% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefactorPriorKnoweldge.R
\name{EquivalentIDs}
\alias{EquivalentIDs}
\title{Find additional potential IDs for  "kegg", "pubchem", "chebi", "hmdb"}
\usage{
EquivalentIDs(
  InputData,
  SettingsInfo = c(InputID = "MetaboliteID"),
  From = "kegg",
  SaveAs_Table = "csv",
  FolderPath = NULL
)
}
\arguments{
\item{InputData}{Dataframe with at least one column with the detected metabolite IDs (one ID per row).}

\item{SettingsInfo}{\emph{Optional: } Column name of metabolite IDs. \strong{Default = list(InputID="MetaboliteID")}}

\item{From}{ID type that is present in your data. Choose between "kegg", "pubchem", "chebi", "hmdb". \strong{Default = "kegg"}}

\item{SaveAs_Table}{\emph{Optional: } File types for the analysis results are: "csv", "xlsx", "txt". \strong{Default = "csv"}}

\item{FolderPath}{{Optional:} Path to the folder the results should be saved at. \strong{Default = NULL}}
}
\value{
Input DF with additional column including potential additional IDs.
}
\description{
Find additional potential IDs for  "kegg", "pubchem", "chebi", "hmdb"
}
\examples{
DetectedIDs <- MetaProViz::ToyData(Data="Cells_MetaData")\%>\% tibble::rownames_to_column("TrivialName")\%>\%tidyr::drop_na()
Res <- MetaProViz::EquivalentIDs(InputData= DetectedIDs, SettingsInfo = c(InputID="KEGG.ID"), From = "kegg")

}
\keyword{Find}
\keyword{IDs}
\keyword{additional}
\keyword{for}
\keyword{identifier}
\keyword{metabolite}
\keyword{one}
\keyword{potential}
