% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyData.R
\docType{data}
\name{gaude_pathways}
\alias{gaude_pathways}
\title{gaude_pathways}
\format{
Data frame with columns: term, gene
}
\source{
Omnipath
}
\usage{
gaude_pathways
}
\description{
gaude_pathways
}
\examples{
data(gaude_pathways)
head(gaude_pathways)
}
\keyword{datasets}
