% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyData.R
\docType{data}
\name{hallmarks}
\alias{hallmarks}
\title{hallmarks}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 7322 rows and 2 columns.
}
\usage{
hallmarks
}
\description{
hallmarks
}
\examples{
data(hallmarks)
head(hallmarks)

}
\keyword{datasets}
