% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPriorKnoweldge.R
\name{InspectID}
\alias{InspectID}
\title{Inspect ID}
\usage{
InspectID(
  Input_DataFrame,
  SettingsInfo = list(OriginalIDcolumn = "MetaboliteID", TranslatedCollapsedIDcolumn =
    "chebi_collapsed", Pathway = "term")
)
}
\value{
Two data frames, the first of which is a summary of the mapping from Original to Translated, and the second of which is the reverse, from Translated to Original, with counts per unique ID and pathway.
}
\description{
Inspect how well IDs map from the translated format (e.g. PubChem) to the original data format (e.g. KEGG), in terms of direct mapping, or one-to-many relationships.
}
