% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{Load_KEGG}
\alias{Load_KEGG}
\title{MCA regulatory rules Import}
\usage{
Load_KEGG()
}
\value{
A data frame containing the KEGG pathways for ORA.
}
\description{
Import and process .csv file to create toy data.
}
\details{
Imports KEGG pathways into the environment
}
