% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPriorKnoweldge.R
\name{LoadGaude}
\alias{LoadGaude}
\title{Prior Knowledge Import}
\usage{
LoadGaude()
}
\value{
A data frame containing the Prior Knowledge.
}
\description{
Import and process .csv file to create Prior Knowledge.
}
