% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPriorKnoweldge.R
\name{LoadHallmarks}
\alias{LoadHallmarks}
\title{Prior Knowledge Import}
\usage{
LoadHallmarks()
}
\value{
A data frame containing the Prior Knowledge.
}
\description{
Import and process .csv file to create Prior Knowledge.
}
