% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPriorKnoweldge.R
\name{LoadKEGG}
\alias{LoadKEGG}
\title{KEGG}
\usage{
LoadKEGG()
}
\value{
A data frame containing the KEGG pathways for ORA.
}
\description{
Import and process KEGG.
}
\details{
Imports KEGG pathways into the environment
}
\examples{
KEGG_Pathways <- MetaProViz::LoadKEGG()

}
