% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPriorKnoweldge.R
\name{LoadRAMP}
\alias{LoadRAMP}
\title{Prior Knowledge Import}
\usage{
LoadRAMP(version = "2.5.4", SaveAs_Table = "csv", FolderPath = NULL)
}
\arguments{
\item{version}{\emph{Optional: } Version of the RaMP database loaded from OmniPathR. \strong{default: "2.5.4"}}

\item{SaveAs_Table}{\emph{Optional: } File types for the analysis results are: "csv", "xlsx", "txt". \strong{Default = "csv"}}

\item{FolderPath}{{Optional:} String which is added to the resulting folder name \strong{default: NULL}}
}
\value{
A data frame containing the Prior Knowledge.
}
\description{
Import and process file to create Prior Knowledge.
}
\examples{
ChemicalClass <- LoadRAMP()

}
