% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPriorKnowledge.R
\name{make_gene_metab_set}
\alias{make_gene_metab_set}
\title{Create metabolite sets from existing genesets}
\usage{
make_gene_metab_set(
  input_pk,
  metadata_info = c(Target = "gene"),
  pk_name = NULL,
  save_table = "csv",
  path = NULL
)
}
\arguments{
\item{input_pk}{dataframe with two columns for source (=term) and Target (=gene), e.g. Hallmarks.}

\item{metadata_info}{\emph{Optional: }  Column name of Target in input_pk. \strong{Default = c(Target="gene")}}

\item{pk_name}{\emph{Optional: } Name of the prior knowledge resource. \strong{default: NULL}}

\item{save_table}{\emph{Optional: } File types for the analysis results are: "csv", "xlsx", "txt". \strong{Default = "csv"}}

\item{path}{{Optional:} String which is added to the resulting folder name \strong{default: NULL}}
}
\value{
List of two data frames: "GeneMetabSet" and "MetabSet".
}
\description{
Gene to metabolite translation is based on mappings in Recon-3D (cosmosR).
}
\examples{
make_gene_metab_set(hallmarks)

}
