% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefactorPriorKnoweldge.R
\name{MappingAmbiguity}
\alias{MappingAmbiguity}
\title{Create Mapping Ambiguities between two ID types}
\usage{
MappingAmbiguity(
  InputData,
  From,
  To,
  GroupingVariable = NULL,
  Summary = FALSE,
  SaveAs_Table = "csv",
  FolderPath = NULL
)
}
\arguments{
\item{InputData}{Translated DF from MetaProViz::TranslateID reults or Dataframe with at least one column with the target metabolite ID and another MetaboliteID type. One of the IDs can only have one ID per row, the other ID can be either separated by comma or a list. Optional: add other columns such as source (e.g. term).}

\item{From}{Column name of the secondary or translated metabolite identifier in InputData. Here can be multiple IDs per row either separated by comma " ," or a list of IDs.}

\item{To}{Column name of original metabolite identifier in InputData. Here should only have one ID per row.}

\item{GroupingVariable}{\emph{Optional: } If NULL no groups are used. If TRUE provide column name in InputData containing the GroupingVariable and features are grouped. \strong{Default = NULL}}

\item{Summary}{\emph{Optional: } If TRUE a long summary tables are created. \strong{Default = FALSE}}

\item{SaveAs_Table}{\emph{Optional: } File types for the analysis results are: "csv", "xlsx", "txt". \strong{Default = "csv"}}

\item{FolderPath}{{Optional:} Path to the folder the results should be saved at. \strong{Default = NULL}}
}
\value{
List with at least 4 DFs: 1-3) From-to-To: 1. MappingIssues, 2. MappingIssues Summary, 3. Long summary (If Summary=TRUE) & 4-6) To-to-From: 4. MappingIssues, 5. MappingIssues Summary, 6. Long summary (If Summary=TRUE) & 7) Combined summary table (If Summary=TRUE)
}
\description{
Create Mapping Ambiguities between two ID types
}
\examples{
KEGG_Pathways <- LoadKEGG()
InputDF <- TranslateID(InputData= KEGG_Pathways, 
    SettingsInfo = c(InputID="MetaboliteID", GroupingVariable="term"), 
    From = c("kegg"), To = c("pubchem"))[["TranslatedDF"]]
Res <- MappingAmbiguity(InputData = InputDF, From = "MetaboliteID", 
    To = "pubchem", GroupingVariable = "term", Summary = TRUE)

}
\keyword{Mapping}
\keyword{ambiguity}
