% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboliteClusteringAnalysis.R
\name{MCA}
\alias{MCA}
\title{Metabolomics pre-processing pipeline}
\usage{
MCA(
  Input_data,
  Input_data2,
  Output_Name = "",
  Condition1,
  Condition2,
  pCutoff = 0.05,
  FCcutoff = 0.5,
  test = "p.adj",
  plot = TRUE,
  CoRe = FALSE,
  Save_as = "svg"
)
}
\arguments{
\item{Input_data}{Dataframe which contains metabolites in rows and Log fold changes, pvalues and padjusted values in columns.}

\item{Input_data2}{Dataframe same as Input_data for another comparison.}

\item{Output_Name}{String which is added to the output files of the Metabolic Clusters.}

\item{Condition1}{String which contains the name of the first condition.}

\item{Condition2}{String which contains the name of the second condition.}

\item{pCutoff}{Number of the desired p value cutoff for assessing significance. \strong{Default = 0.05}}

\item{FCcutoff}{Number of the desired log fold change cutoff for assessing significance. \strong{Default = 0.5}}

\item{test}{String which selects pvalue or padj for significance.  \strong{Default = padj}}

\item{plot}{\emph{Optional: } TRUE or FALSE, if TRUE Volcano plot is saved as an overview of the results. \strong{Default = TRUE}}

\item{CoRe}{\emph{Optional: } TRUE or FALSE for whether a Consumption/Release  input is used. \strong{Default = FALSE}}

\item{Save_as}{\emph{Optional: } Select the file type of output plots. Options are svg or pdf. \strong{Default = svg}}
}
\description{
Metabolomics pre-processing pipeline
}
\author{
Prymidis Dimitrios, Schmidt Christina
Date: "2022-10-28"

This script allows you to perform metabolite clustering analysis
}
\keyword{Clusters,}
\keyword{Metabolic}
