% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyData.R
\docType{data}
\name{mca_2cond}
\alias{mca_2cond}
\title{mca_2cond}
\format{
Columns include Intra, core, core_Detection including state entries (e.g. up, down, etc.) and the Regulator Clustering
columns (RG1-RG3)
}
\source{
Schmidt et al, MetaProViz: METabolomics pre-PRocessing, functiOnal analysis and VIZualisation version 2.1.7, GitHub 2025.
}
\usage{
mca_2cond
}
\description{
Manually curated table defining the flow of information of the two condition biological regulatory clusters Regulatory labels
from the different grouping methods.
}
\keyword{datasets}
