% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyData.R
\docType{data}
\name{mca_core_rules}
\alias{mca_core_rules}
\title{mca_core_rules}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 108 rows and 7 columns.
}
\usage{
mca_core_rules
}
\description{
Manually curated table defining the flow of information of the
Conusuption-Release and Intracellular metabolomics biological regulatory
clusters Regulatory labels from the different grouping methods. columns
(RG1-RG3)
}
\examples{
data(mca_core_rules)
head(mca_core_rules)

}
\keyword{datasets}
