% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyData.R
\docType{data}
\name{mca_twocond_rules}
\alias{mca_twocond_rules}
\title{mca_twocond_rules}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 36 rows and 5 columns.
}
\usage{
mca_twocond_rules
}
\description{
Manually curated table defining the flow of information of the two condition
biological regulatory clusters Regulatory labels from the different grouping
methods. columns (RG1-RG3)
}
\examples{
data(mca_twocond_rules)
head(mca_twocond_rules)

}
\keyword{datasets}
