% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyData.R
\docType{data}
\name{medium_raw}
\alias{medium_raw}
\title{medium_raw}
\format{
Columns include Conditions (=Cell lines, blanks),Biological_Replicate, GrowthFactor and
a numeric column for each measured metabolite (raw data)
}
\source{
Sciacovelli & Dugourd et. al., Dynamic partitioning of branched-chain amino acids-derived
nitrogen supports renal cancer progression , Nature Communications 2022, \doi{10.1038/s41467-022-35036-4}
}
\usage{
medium_raw
}
\description{
Metabolomics workbench project PR001418, study ST002226 where we exported integrated raw
peak values of intracellular metabolomics of HK2 and cccRCC cell lines 786-O, 786-M1A,
786-M2A, OS-RC-2, OS-LM1 and RFX-631.
}
\examples{
data(medium_raw)
head(medium_raw)
}
\keyword{datasets}
