% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyData.R
\docType{data}
\name{medium_raw_se}
\alias{medium_raw_se}
\title{medium_raw_se}
\format{
An object of class \code{SummarizedExperiment} with 73 rows and 44 columns.
}
\usage{
medium_raw_se
}
\description{
Metabolomics workbench project PR001418, study ST002226 where we exported
integrated raw peak values of intracellular metabolomics of HK2 and cccRCC
cell lines 786-O, 786-M1A, 786-M2A, OS-RC-2, OS-LM1 and RFX-631 converted
into an se object. -\code{Conditions}: Character vector indicating cell
line identity -\code{Biological_Replicates}: Integer replicate number for
biological replicates -\code{GrowthFactor}: Character vector indicating
growth factor condition nitrogen supports renal cancer progression,
Nature Communications 2022, \doi{10.1038/s41467-022-35036-4}
}
\examples{
data(medium_raw_se)
head(medium_raw_se)

}
\keyword{datasets}
