% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaDataAnalysis.R
\name{Meta_Corr}
\alias{Meta_Corr}
\title{This script allows you to perform}
\usage{
Meta_Corr(
  Input_SettingsInfo = NULL,
  Input_SettingsFile,
  Input_data,
  InputSettingsFile_Feature,
  Input_SettingsInfo_Feature,
  Scaling = TRUE,
  percentage = 0.1,
  OutputName = "",
  Save_as_Results = "svg",
  Folder_Name = NULL,
  path = NULL
)
}
\arguments{
\item{Input_SettingsInfo}{\emph{Optional: } NULL or Named vector including at least one of those three information : c(color="ColumnName_Plot_SettingsFile", shape= "ColumnName_Plot_SettingsFile"). \strong{Default = NULL}}

\item{Input_data}{DF with unique sample identifiers as row names and metabolite numerical values in columns with metabolite identifiers as column names. Use NA for metabolites that were not detected. includes experimental design and outlier column.}

\item{Input_SettingsInfo_Feature}{\emph{Optional: } DF which contains information about the features (=metabolites)}

\item{Scaling}{\emph{Optional: } TRUE or FALSE for whether a data scaling is used \strong{Default = TRUE}}

\item{Save_as_Results}{\emph{Optional: } Select the file type of output plots. Options are svg, png, pdf or NULL. \strong{Default = svg}}

\item{Folder_Name}{{Optional:} String which is added to the resulting folder name \strong(Default = NULL)}

\item{Input_SettingsFile_Feature}{\emph{Optional: } DF which contains information about the samples, which will be combined with your input data based on the unique sample identifiers used as rownames. Column "Conditions" with information about the sample conditions (e.g. "N" and "T" or "Normal" and "Tumor"), can be used for feature filtering and colour coding in the PCA. Column "AnalyticalReplicate" including numerical values, defines technical repetitions of measurements, which will be summarised. Column "BiologicalReplicates" including numerical values. Please use the following names: "Conditions", "Biological_Replicates", "Analytical_Replicates".\strong{Default = NULL}}

\item{OutputPlotName}{\emph{Optional: } String which is added to the output files of the PCA \strong{Default = ""}}
}
\description{
This script allows you to perform
}
\keyword{PCA}
