% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperLog.R
\name{MetaProViz_log}
\alias{MetaProViz_log}
\title{Browse the current MetaProViz log file}
\usage{
MetaProViz_log()
}
\value{
Returns \code{NULL}.
}
\description{
Browse the current MetaProViz log file
}
\examples{
\dontrun{
MetaProViz_log()
# then you can browse the log file, and exit with `q`
}

}
\seealso{
\code{\link{MetaProViz_logfile}}
}
