% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{metaproviz_logfile}
\alias{metaproviz_logfile}
\title{Path to the current MetaProViz log file}
\usage{
metaproviz_logfile()
}
\value{
Character: path to the current logfile, or \code{NULL} if no
logfile is available.
}
\description{
Path to the current MetaProViz log file
}
\examples{
metaproviz_logfile()
# [1] "/home/denes/metaproviz/metaproviz-log/metaproviz-20210309-1642.log"

}
\seealso{
\code{\link{metaproviz_log}}
}
