% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperOptions.R
\name{metaproviz_reset_config}
\alias{metaproviz_reset_config}
\title{Restore the built-in default values of all config parameters of MetaProViz}
\usage{
metaproviz_reset_config(save = NULL, reset_all = FALSE)
}
\arguments{
\item{save}{If a path, the restored config will be also saved
to this file. If TRUE, the config will be saved to the current default
config path (see \code{\link{metaproviz_config_path}}).}

\item{reset_all}{Reset to their defaults also the options already set in
the R options.}
}
\value{
The config as a list.
}
\description{
Restore the built-in default values of all config parameters of MetaProViz
}
\examples{
\dontrun{
# restore the defaults and write them to the default config file:
metaproviz_reset_config()
metaproviz_save_config()
}

}
\seealso{
\code{\link{metaproviz_load_config}, \link{metaproviz_save_config}}
}
