% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPriorKnowledge.R
\name{metsigdb_kegg}
\alias{metsigdb_kegg}
\title{KEGG pathways}
\usage{
metsigdb_kegg(exclude_metabolites = "all")
}
\arguments{
\item{exclude_metabolites}{Optional metabolite classes to exclude:
NULL (exclude nothing), "all" (default), or any combination of
c("ions", "small_molecules", "xenobiotics", "atoms").}
}
\value{
A data frame containing the KEGG pathways suitable for ORA.
}
\description{
KEGG pathways
}
\examples{
metsigdb_kegg()

}
