% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPriorKnowledge.R, R/devel_metsigdb_new.R
\name{metsigdb_reactome}
\alias{metsigdb_reactome}
\title{Retrieve Reactome metabolite sets suitable for ORA.}
\usage{
metsigdb_reactome(species = "Homo sapiens")

metsigdb_reactome(species = "Homo sapiens")
}
\arguments{
\item{species}{String. Optionally specify pathways to query from a species
via full name or three letter code. Default = "Homo sapiens". NULL for all species.}

\item{pathway_ids}{String vector. Optionally provide pathway_ids to query. Default NULL to query all pathways.}

\item{out_path}{String. Optionally save results as csv into out_path. Default NULL.}
}
\value{
A tibble in long format containing one row per metabolite for the Reactome pathways.

A tibble in long format containing one row per metabolite for the Reactome pathways.
}
\description{
Queries the OmniPath resource through OmniPathR to obtail Reactome pathway
level metabolite sets.

Queries the OmniPath resource through OmniPathR to obtail Reactome pathway
level metabolite sets.
}
\examples{
\dontrun{
df <- metsigdb_reactome()
head(df)
}

\dontrun{
df <- metsigdb_reactome()
head(df)
}

}
