% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing.R
\name{mvi_imputation}
\alias{mvi_imputation}
\title{Missing Value Imputation using half minimum value}
\usage{
mvi_imputation(
  data,
  metadata_sample,
  metadata_info,
  core = FALSE,
  mvi_percentage = 50
)
}
\arguments{
\item{data}{DF which contains unique sample identifiers as row names and metabolite
numerical values in columns with metabolite identifiers as column names.
Use NA for metabolites that were not detected and consider converting
any zeros to NA unless they are true zeros.}

\item{metadata_sample}{DF which contains information about the samples, which will be combined
with the input data based on the unique sample identifiers used as
rownames.}

\item{metadata_info}{Named vector containing the information about the names of the
experimental parameters. c(Conditions="ColumnName_Plot_SettingsFile",
Biological_Replicates="ColumnName_Plot_SettingsFile", core_media =
"Columnname_Input_SettingsFile"). Column "Conditions" with information
about the sample conditions, Column "BiologicalReplicates" including
numerical values and Column "Columnname_Input_SettingsFile" is used to
specify the name of the media controls in the Conditions.}

\item{core}{\emph{Optional: } If TRUE, a consumption-release experiment has been
performed. Should not be normalised to media blank. Provide information
about control media sample names via metadata_info "core_media"
samples.\strong{Default = FALSE}}

\item{mvi_percentage}{\emph{Optional: } percentage 0-100 of imputed value based on the minimum
value. \strong{Default = 50}}
}
\value{
DF with imputed values

This function can be used on raw or already filtered data, provided sample
metadata and condition information are available.
}
\description{
Missing Value Imputation using half minimum value
}
\examples{
data(intracell_raw)
Intra <- intracell_raw \%>\% tibble::column_to_rownames("Code")
Res <- mvi_imputation(
    data =
        Intra[-c(49:58), -c(1:3)] \%>\%
        dplyr::mutate_all(
            ~ ifelse(grepl("^0*(\\\\.0*)?$", as.character(.)), NA, .)
        ),
    metadata_sample = Intra[-c(49:58), c(1:3)],
    metadata_info = c(
        Conditions = "Conditions",
        Biological_Replicates = "Biological_Replicates"
    )
)

}
