% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{Pool_Estimation}
\alias{Pool_Estimation}
\title{Description}
\usage{
Pool_Estimation(
  Input_data,
  Input_SettingsFile = NULL,
  Input_SettingsInfo = NULL,
  Unstable_feature_remove = FALSE,
  Therhold_cv = 1,
  Threshold_SEMean = 0.1,
  Threshold_SEMedian = 0.1
)
}
\arguments{
\item{Input_data}{DF which contains unique sample identifiers as row names and metabolite numerical values in columns with metabolite identifiers as column names. Use NA for metabolites that were not detected. Can be either a full dataset or a dataset with just the pooled samples.}

\item{Input_SettingsFile}{\emph{Optional: } DF which contains information about the samples, which will be combined with your input data based on the unique sample identifiers used as rownames when a full dataset is inserted as Input_data. Column "Conditions" with information about the sample conditions (e.g. "N" and "T" or "Normal" and "Tumor"), exist.\strong{Default = NULL}}

\item{Input_SettingsInfo}{\emph{Optional: } NULL or Named vector including the Pooled_Sample information (Name of the pooled samples in the Conditions in the Input_SettingsFile)  : c(Conditions="Pooled_Samples). \strong{Default = NULL}}

\item{Unstable_feature_remove}{\emph{Optional: }  Parameter to automatically remove unstable(high variance) metabolites from the dataset. Used only when a full dataset is used as Input_data. \strong{Default = FALSE}}

\item{Therhold_cv}{\emph{Optional: } Filtering threshold for high variance metabolites using the Coefficient of Variation. \strong{Default = 1}}

\item{Threshold_SEMean}{\emph{Optional: } Filtering threshold for high variance metabolites using the Standard Error of the Mean ration to the Mean. \strong{Default = 0.1}}

\item{Threshold_SEMedian}{\emph{Optional: } Filtering threshold for high variance metabolites using the Standard Error of the Median ration to the Mean. \strong{Default = 0.1}}
}
\description{
Description
}
\keyword{Coefficient}
\keyword{Error}
\keyword{Mean,}
\keyword{Standard}
\keyword{Variation,}
\keyword{high}
\keyword{metabolites}
\keyword{of}
\keyword{the}
\keyword{variance}
