% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefactorPriorKnoweldge.R
\name{PossibleID}
\alias{PossibleID}
\title{Deal with detected input features.}
\usage{
PossibleID(
  InputData,
  SettingsInfo = c(InputID = "MetaboliteID", GroupingVariable = "term")
)
}
\arguments{
\item{InputData}{Dataframe with at least one column with the target (e.g. metabolite), you can add other columns such as source (e.g. term)}

\item{SettingsInfo}{}
}
\description{
Deal with detected input features.
}
