% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{Preprocessing}
\alias{Preprocessing}
\title{Applies 80\%-filtering rule, total-ion count normalisation, missing value imputation and HotellingT2 outlier detection}
\usage{
Preprocessing(
  Input_data,
  Input_SettingsFile,
  Input_SettingsInfo,
  Feature_Filtering = "Modified",
  Feature_Filt_Value = 0.8,
  TIC_Normalization = TRUE,
  MVI = TRUE,
  MVI_Percentage = 50,
  HotellinsConfidence = 0.99,
  CoRe = FALSE,
  ExportQCPlots = TRUE,
  Save_as_Plot = "svg",
  Plot = TRUE,
  Folder_Name = NULL
)
}
\arguments{
\item{Input_data}{DF which contains unique sample identifiers as row names and metabolite numerical values in columns with metabolite identifiers as column names. Use NA for metabolites that were not detected.}

\item{Input_SettingsFile}{DF which contains information about the samples, which will be combined with the input data based on the unique sample identifiers used as rownames.}

\item{Input_SettingsInfo}{NULL or Named vector containing the information about the names of the experimental parameters. c(Conditions="ColumnName_Plot_SettingsFile", Biological_Replicates="ColumnName_Plot_SettingsFile"). Column "Conditions" with information about the sample conditions (e.g. "N" and "T" or "Normal" and "Tumor"), can be used for feature filtering and colour coding in the PCA. Column "BiologicalReplicates" including numerical values. For CoRe = TRUE a CoRe_norm_factor = "Columnname_Input_SettingsFile" and CoRe_media = "Columnname_Input_SettingsFile", have to also be added. Column CoRe_norm_factor is used for normalization and CoRe_media is used to specify the name of the media controls in the Conditions.}

\item{Feature_Filtering}{\emph{Optional: }If set to "None" then no feature filtering is performed. If set to "Standard" then it applies the 80\%-filtering rule (Bijlsma S. et al., 2006) on the metabolite features on the whole dataset. If is set to "Modified",filtering is done based on the different conditions, thus a column named "Conditions" must be provided in the Input_SettingsFile input file including the individual conditions you want to apply the filtering to (Yang, J et al., 2015). \strong{Default = Modified}}

\item{Feature_Filt_Value}{\emph{Optional: } Percentage of feature filtering. \strong{Default = 0.8}}

\item{TIC_Normalization}{\emph{Optional: } If TRUE, Total Ion Count normalization is performed. \strong{Default = TRUE}}

\item{MVI}{\emph{Optional: } If TRUE, Missing Value Imputation (MVI) based on half minimum is performed \strong{Default = TRUE}}

\item{MVI_Percentage}{\emph(Optional: ) Percentage (0 to 100)of imputed value based on the minimum value. \strong{Default = 50}}

\item{HotellinsConfidence}{\emph{Optional: } Defines the Confidence of Outlier identification in HotellingT2 test. Must be numeric.\strong{Default = 0.99}}

\item{CoRe}{\emph{Optional: } If TRUE, a consumption-release experiment has been performed and the CoRe value will be calculated. Please consider providing a Normalisation factor column called "CoRe_norm_factor" in your "Input_SettingsFile" DF, where the column "Conditions" matches. The normalisation factor must be a numerical value obtained from growth rate that has been obtained from a growth curve or growth factor that was obtained by the ratio of cell count/protein quantification at the start point to cell count/protein quantification at the end point.. Additionally control media samples have to be available in the "Input" DF and defined as "CoRe_media" samples in the "Conditions" column in the "Input_SettingsFile" DF. \strong{Default = FALSE}}

\item{ExportQCPlots}{\emph{Optional: } Select whether the quality control (QC) plots will be exported. \strong{Default = TRUE}}

\item{Save_as_Plot}{\emph{Optional: } Select the file type of output plots. Options are svg, png, pdf. \strong{Default = svg}}

\item{Plot}{\emph{Optional: } If TRUE prints an overview of resulting plots. \strong{Default = TRUE}}

\item{Folder_Name}{{Optional:} String which is added to the resulting folder name \strong(Default = NULL)}
}
\description{
Applies 80\%-filtering rule, total-ion count normalisation, missing value imputation and HotellingT2 outlier detection
}
\keyword{80%}
\keyword{Count}
\keyword{HotellingT2,}
\keyword{Imputation,}
\keyword{Ion}
\keyword{Missing}
\keyword{PCA,}
\keyword{Total}
\keyword{Value}
\keyword{charts,}
\keyword{control}
\keyword{filtering}
\keyword{multivariate}
\keyword{normalization,}
\keyword{quality}
\keyword{rule,}
