% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing.R
\name{replicate_sum}
\alias{replicate_sum}
\title{Merges the analytical replicates of an experiment}
\usage{
replicate_sum(
  data,
  metadata_sample,
  metadata_info = c(Conditions = "Conditions", Biological_Replicates =
    "Biological_Replicates", Analytical_Replicates = "Analytical_Replicates"),
  save_table = "csv",
  path = NULL
)
}
\arguments{
\item{data}{DF which contains unique sample identifiers as row names and metabolite numerical values in columns with metabolite identifiers as column names. Use NA for metabolites that were not detected.}

\item{metadata_sample}{DF which contains information about the samples Column "Conditions", "Biological_replicates" and "Analytical_Replicates has to exist.}

\item{metadata_info}{\emph{Optional: } Named vector including the Conditions and Replicates information: c(Conditions="ColumnNameConditions", Biological_Replicates="ColumnName_metadata_sample", Analytical_Replicates="ColumnName_metadata_sample").\strong{Default = NULL}}

\item{save_table}{\emph{Optional: } File types for the analysis results are: "csv", "xlsx", "txt", ot NULL \strong{default: "csv"}}

\item{path}{\emph{Optional:} Path to the folder the results should be saved at. \strong{default: NULL}}
}
\value{
DF with the merged analytical replicates
}
\description{
Merges the analytical replicates of an experiment
}
\examples{
Intra <- ToyData("IntraCells_Raw")
Res <- replicate_sum(data=Intra[-c(49:58) ,-c(1:3)],
                                metadata_sample=Intra[-c(49:58) , c(1:3)],
                                metadata_info = c(Conditions="Conditions", Biological_Replicates="Biological_Replicates", Analytical_Replicates="Analytical_Replicates"))

}
\keyword{Analytical}
\keyword{Merge}
\keyword{Replicate}
