% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverRepresentationAnalysis.R
\name{StandardORA}
\alias{StandardORA}
\title{Uses enricher to run ORA on the differential metabolites (DM) using a pathway list}
\usage{
StandardORA(
  data,
  metadata_info = c(pvalColumn = "p.adj", percentageColumn = "t.val", PathwayTerm =
    "term", PathwayFeature = "Metabolite"),
  cutoff_stat = 0.05,
  cutoff_percentage = 10,
  input_pathway,
  pathway_name = "",
  min_gssize = 10,
  max_gssize = 1000,
  save_table = "csv",
  path = NULL
)
}
\arguments{
\item{data}{DF with metabolite names/metabolite IDs as row names. Metabolite names/IDs need to match the identifier type (e.g. HMDB IDs) in the input_pathway.}

\item{metadata_info}{\emph{Optional: } Pass ColumnName of the column including parameters to use for cutoff_stat and cutoff_percentage. Also pass ColumnName for input_pathway including term and feature names. (pvalColumn = ColumnName data, percentageColumn= ColumnName data, PathwayTerm= ColumnName input_pathway, PathwayFeature= ColumnName input_pathway) \strong{c(pvalColumn="p.adj", percentageColumn="t.val", PathwayTerm= "term", PathwayFeature= "Metabolite")}}

\item{cutoff_stat}{\emph{Optional: } p-adjusted value cutoff from ORA results. Must be a numeric value. \strong{default: 0.05}}

\item{cutoff_percentage}{\emph{Optional: } percentage cutoff of metabolites that should be considered for ORA. Selects top/Bottom \% of selected percentageColumn, usually t.val or Log2FC \strong{default: 10}}

\item{input_pathway}{DF that must include column "term" with the pathway name, column "Metabolite" with the Metabolite name or ID and column "Description" with pathway description that will be depicted on the plots.}

\item{pathway_name}{\emph{Optional: } Name of the input_pathway used \strong{default: ""}}

\item{min_gssize}{\emph{Optional: } minimum group size in ORA \strong{default: 10}}

\item{max_gssize}{\emph{Optional: } maximum group size in ORA \strong{default: 1000}}

\item{save_table}{\emph{Optional: } File types for the analysis results are: "csv", "xlsx", "txt" \strong{default: "csv"}}

\item{path}{\emph{Optional:} Path to the folder the results should be saved at. \strong{default: NULL}}
}
\value{
Saves results as individual .csv files.
}
\description{
Uses enricher to run ORA on the differential metabolites (DM) using a pathway list
}
