% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing.R
\name{tic_norm}
\alias{tic_norm}
\title{Total ion count normalization}
\usage{
tic_norm(data, metadata_sample, metadata_info, tic = TRUE)
}
\arguments{
\item{data}{DF which contains unique sample identifiers as row names and metabolite
numerical values in columns with metabolite identifiers as column names.
Use NA for metabolites that were not detected and consider converting
any zeros to NA unless they are true zeros.}

\item{metadata_sample}{DF which contains information about the samples, which will be combined
with the input data based on the unique sample identifiers used as
rownames.}

\item{metadata_info}{Named vector containing the information about the names of the
experimental parameters. c(Conditions="ColumnName_Plot_SettingsFile").}

\item{tic}{\emph{Optional: } If TRUE, total Ion Count normalization is performed.
If FALSE, only RLA QC plots are returned. \strong{Default = TRUE}}
}
\value{
List with two elements: DF (including output table) and Plot (including
all plots generated)

This function can be used independently on numeric data with matching sample
metadata. Missing values and zero handling should already be in a reasonable
state for the intended normalization.
}
\description{
Total ion count normalization
}
\examples{
data(intracell_raw)
Intra <- intracell_raw \%>\% tibble::column_to_rownames("Code")
Res <- tic_norm(
    data =
        Intra[-c(49:58), -c(1:3)] \%>\%
        dplyr::mutate_all(
            ~ ifelse(grepl("^0*(\\\\.0*)?$", as.character(.)), NA, .)
        ),
    metadata_sample = Intra[-c(49:58), c(1:3)],
    metadata_info = c(Conditions = "Conditions")
)

}
