% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyData.R
\docType{data}
\name{tissue_meta}
\alias{tissue_meta}
\title{Tissue_Metadata}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 877 rows and 11 columns.
}
\usage{
tissue_meta
}
\description{
In Hakimi et. al. metabolites were assigned to metabolite IDs, pathways,
platform, mass and other fetaure metainformation. row names being metabolite
trivial names. \doi{10.1016/j.ccell.2015.12.004}
}
\examples{
data(tissue_meta)
head(tissue_meta)

}
\keyword{datasets}
