% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyData.R
\docType{data}
\name{tissue_norm_se}
\alias{tissue_norm_se}
\title{tissue_norm_se}
\format{
An object of class \code{SummarizedExperiment} with 570 rows and 276 columns.
}
\usage{
tissue_norm_se
}
\description{
This is median normalised data from the supplementary table 2 of Hakimi et
al with metabolomic profiling on 138 matched clear cell renal cell carcinoma
(ccRCC)/normal tissue pairs. coldata include patient metadata (e.g. age,
gender, sage, etc.) \doi{10.1016/j.ccell.2015.12.004}
}
\examples{
data(tissue_norm_se)
head(tissue_norm_se)

}
\keyword{datasets}
