% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyData.R
\docType{data}
\name{tissue_tvn_rnaseq}
\alias{tissue_tvn_rnaseq}
\title{tissue_tvn_rnaseq}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 29283 rows and 10 columns.
}
\usage{
tissue_tvn_rnaseq
}
\description{
The processed transcriptomics data was downloaded from the supplementary
table 3 of Mora & Schmidt et. al., which used the study from Clark et. al.
under Proteomics data Commons PDC000127. on their regulation regulation in
renal cancer, Genome Medicine 2024, \doi{10.1186/s13073-024-01415-3} Clark
et. al, Integrated proteogenomic characterization of clear cell renal cell
carcinoma, Cell 2019, \doi{10.1016/j.cell.2019.10.007}
}
\examples{
data(tissue_tvn_rnaseq)
head(tissue_tvn_rnaseq)

}
\keyword{datasets}
