% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{toy_data}
\alias{toy_data}
\title{Toy Data Import}
\usage{
toy_data()
}
\value{
A data frame containing the toy data.
}
\description{
Import and process .csv file to create toy data.
}
\details{
Imports toy data into environment
}
