% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{ToyData}
\alias{ToyData}
\title{Access built-in example data}
\usage{
ToyData(name)
}
\arguments{
\item{name}{Character: name of a built-in dataset; "Intra",
"Intra_DMA", "Media" or "Pathways" are availale.}
}
\value{
A data frame containing the toy data.
}
\description{
Import and process .csv file to create toy data.
}
\examples{
intra <- ToyData("Intra")

}
