% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizStackedbar.R
\name{viz_stackedbar}
\alias{viz_stackedbar}
\title{Generate Stacked Bar Plot, e.g. for PK Coverage}
\usage{
viz_stackedbar(
  data,
  group_col,
  fill_col,
  fill_values,
  fill_labels,
  plot_title,
  x_label = "Frequency",
  y_label = NULL,
  legend_position = c(0.95, 0.05)
)
}
\arguments{
\item{data}{A data frame containing the data to be plotted.}

\item{group_col}{A string specifying the name of the column to group the data by.}

\item{fill_col}{A string specifying the name of the column to use for fill aesthetics.}

\item{fill_values}{A vector of color values to be used for the fill aesthetic.}

\item{fill_labels}{A vector of labels corresponding to the fill levels for the legend.}

\item{plot_title}{A string specifying the title of the plot.}

\item{x_label}{A string for the x-axis label. Defaults to "Frequency".}

\item{y_label}{A string for the y-axis label. If \code{NULL}, the value of \code{group_col} is used.}

\item{legend_position}{A numeric vector of length 2 specifying the (x, y) position of the legend.
Defaults to \code{c(0.95, 0.05)}.}
}
\value{
A \code{ggplot} object representing the stacked bar plot.
}
\description{
This function creates a stacked bar plot, for example to visualize Prior Knowledge (PK) coverage.
The plot groups data by a specified column and fills the bars according to another column,
allowing you to inspect the distribution of match statuses (or any categorical variable).
}
