% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizUpset.R
\name{viz_upset}
\alias{viz_upset}
\title{Generate Complex Upset Plot for PK Coverage (Optional Class Grouping)}
\usage{
viz_upset(
  df,
  class_col = NULL,
  intersect_cols = c("LIMID", "HMDB", "CHEBI", "None"),
  plot_title = "Metabolite IDs",
  palette_type = c("viridis", "polychrome"),
  output_file = NULL,
  width = 14,
  height = 8,
  dpi = 300,
  max_legend_terms = 20,
  save_plot = "svg",
  print_plot = TRUE,
  path = NULL
)
}
\arguments{
\item{df}{A data frame containing the data to be plotted.}

\item{class_col}{\emph{Optional: } An optional string specifying the name of the column in \code{df} that represents the class
of each observation. This column is coerced to a factor if provided. \strong{Default = NULL}}

\item{intersect_cols}{\emph{Optional: } A character vector specifying the names of the columns in \code{df} to be used for generating intersections.
\strong{Default = c("LIMID", "HMDB", "CHEBI", "None")}.}

\item{plot_title}{\emph{Optional: } A string specifying the title of the plot. \strong{Default = "Metabolite IDs"}.}

\item{palette_type}{\emph{Optional: } A string specifying the color palette to use for the fill aesthetic when \code{class_col} is provided.
Options are \code{"viridis"} and \code{"polychrome"}. \strong{Default = c("viridis", "polychrome")}}

\item{output_file}{\emph{Optional: } An optional string specifying the file path to save the plot. If \code{NULL}, the plot is not saved. \strong{Default = NULL}}

\item{width}{\emph{Optional: } Numeric value specifying the width of the saved plot (if \code{output_file} is provided). \strong{Default = 14}.}

\item{height}{\emph{Optional: } Numeric value specifying the height of the saved plot (if \code{output_file} is provided). \strong{Default = 8}.}

\item{dpi}{\emph{Optional: } Numeric value specifying the resolution (dots per inch) of the saved plot (if \code{output_file} is provided). \strong{Default = 300}.}

\item{max_legend_terms}{\emph{Optional: } Numeric value specifying the maximum number of unique terms in \code{class_col}
for which the legend should be displayed. If the number of levels exceeds this value,
the legend will be hidden. Ignored if \code{class_col} is \code{NULL}. \strong{Default = 20}.}

\item{save_plot}{\emph{Optional: } Select the file type of output plots. Options are svg, png, pdf or NULL. \strong{Default = svg}}

\item{print_plot}{\emph{Optional: } TRUE or FALSE, if TRUE Volcano plot is saved as an overview of the results. \strong{Default = TRUE}}

\item{path}{\emph{Optional:} Path to the folder the results should be saved at. \strong{Default= NULL}}
}
\value{
A \code{ggplot} object representing the generated upset plot.
}
\description{
This function creates a complex upset plot.
It is mostly intended to visualize Prior Knowledge (PK) coverage by displaying the intersections among
a set of metabolite ID columns, but its use could be extended too. If a class column is provided,
the data is grouped by that column and a color palette ("viridis" or "polychrome") is used to represent
the class levels, along with a corresponding legend (which can be hidden if there are too many unique classes).
If no class column is provided (\code{class_col = NULL}), a basic upset plot is generated.
}
\examples{
# example code

}
\keyword{complex}
\keyword{plot,}
\keyword{upset}
