% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{VizAlluvial}
\alias{VizAlluvial}
\title{Notes; Check the alluvial in the Metabolic Clusters. This one is a bit old but kept here to return to if nedded}
\usage{
VizAlluvial(
  Input_data1,
  Input_data2,
  Output_Name = "Metabolic_Clusters_Output_Condition1-versus-Condition2",
  Condition1,
  Condition2,
  pCutoff = 0.05,
  FCcutoff = 0.5,
  test = "p.adj",
  plot_column_names = c("class", "MetaboliteChange_Significant", "Overall_Change",
    "Metabolite"),
  safe_colorblind_palette = c("#88CCEE", "#DDCC77", "#661100", "#332288", "#AA4499",
    "#999933", "#44AA99", "#882255", "#6699CC", "#117733", "#888888", "#CC6677", "#FFF",
    "#000"),
  plot_color_variable = "Overall_Change",
  plot_color_remove_variable = "SameDirection_NoChange",
  Save_as = pdf
)
}
\description{
Notes; Check the alluvial in the Metabolic Clusters. This one is a bit old but kept here to return to if nedded
}
