% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizLolipop.R
\name{VizLolipop}
\alias{VizLolipop}
\title{This script allows you to perform different data visualizations using the results of the MetaProViz analysis}
\usage{
VizLolipop(
  Plot_Settings = "Standard",
  Plot_SettingsInfo = NULL,
  Plot_SettingsFile = NULL,
  Input_data,
  AdditionalInput_data = NULL,
  x = "Log2FC",
  y = "Metabolite",
  OutputPlotName = "",
  Comparison_name = c(Input_data = "Cond1", AdditionalInput_data = "Cond2"),
  Subtitle = "",
  Theme = NULL,
  Save_as_Plot = "svg",
  parameter_size = "Reverse"
)
}
\arguments{
\item{Plot_Settings}{\emph{Optional: } Choose between "Standard" (Input_data), "Compare" (plot two comparisons together Input_data and Input_data2) or "PEA" (Pathway Enrichment Analysis) \strong{Default = "Standard"}}

\item{Plot_SettingsInfo}{\emph{Optional: } NULL or Named vector including at least one of those three information for Plot_Settings="Standard" or "Compare": c(color="ColumnName_Plot_SettingsFile", shape= "ColumnName_Plot_SettingsFile", individual="ColumnName_Plot_SettingsFile"). For Plot_Settings="PEA" a named vector with c(PEA_Pathway="ColumnNameAdditionalInput_data", PEA_score="ColumnNameAdditionalInput_data", PEA_stat= "ColumnNameAdditionalInput_data", individual="Plot_SettingsFile), optionally you can additionally include c(color="ColumnName_Plot_SettingsFile", shape= "ColumnName_Plot_SettingsFile").\strong{Default = NULL}}

\item{Plot_SettingsFile}{\emph{Optional: } DF with column "Metabolite" including the Metabolite names (needs to match Metabolite names of Input_data) and other columns with required PlotSettingInfo. \strong{Default = NULL}}

\item{Input_data}{DF with column "Metabolite" including the Metabolite names, Log2FC, pvalue/padjusted values. Can also include additional columns with metadata usable for Plot_Setting_Info. Multiple Input data frames can be added using a list ie. list(df1, df2, df3)}

\item{OutputPlotName}{\emph{Optional: } String which is added to the output files of the plot. \strong{Default = ""}}

\item{Comparison_name}{\emph{Optional: } List including those information about the datasets that are compared on the plots when choosing Plot_Settings= "Compare". \strong{Default = list("Cond1", "Cond2")}}

\item{Subtitle}{\emph{Optional: } \strong{Default = ""}}

\item{Theme}{\emph{Optional: } Selection of theme for plot. \strong{Default = NULL}}

\item{Save_as_Plot}{\emph{Optional: } Select the file type of output plots. Options are svg, pdf or png. \strong{Default = "svg"}}

\item{test}{\emph{Optional: } String which selects pvalue or padj for significance. \strong{Default = padj}}

\item{pCutoff}{\emph{Optional: } Number of the desired p value cutoff for assessing significance. \strong{Default = 0.05}}

\item{FCcutoff}{\emph{Optional: } Number of the desired log fold change cutoff for assessing significance. \strong{Default = 0.5}}

\item{Legend}{\emph{Optional: } Legend=="Pie" will plot a PieChart as the legend for color or Legend="Standard, plot the standard legend for color. \strong{Default = "Standard"}}
}
\description{
This script allows you to perform different data visualizations using the results of the MetaProViz analysis
}
\keyword{Lolipop}
\keyword{pathways}
\keyword{plot,}
