% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyData.R
\docType{data}
\name{alanine_pathways}
\alias{alanine_pathways}
\title{alanine_pathways}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 204 rows and 5 columns.
}
\usage{
alanine_pathways
}
\description{
Manually curated table for the amino acid alanine toshowcase pathways (wiki,
reactome, etc.) and alanine IDs (chebi, hmdb, etc.) included in those
pathways
}
\examples{
data(alanine_pathways)
head(alanine_pathways)

}
\keyword{datasets}
