% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyData.R
\docType{data}
\name{biocrates_features}
\alias{biocrates_features}
\title{biocrates_features}
\format{
Columns include metabolite trivial name, metabolite class, metabolite IDs (HMDB, KEGG, etc.), metabolite structural
information (INCHI, Key, etc.).
}
\source{
Biocrates MxP® Quant 500 XL kit, https://biocrates.com/mxp-quant-500-xl/
}
\usage{
biocrates_features
}
\description{
Biocrates kit feature information of the "MxP Quant 500 XL kit" that covers more than 1,000 metabolites with biochemical
class information and the exported different metabolite IDs (HMDB, KEGG, etc.).
}
\keyword{datasets}
