% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyData.R
\docType{data}
\name{cellular_meta}
\alias{cellular_meta}
\title{cellular_meta}
\format{
Columns include Metabolite IDs (HMDB, KEGG), Main metabolic pathway with row names being
metabolite trivial names.
}
\source{
Sciacovelli & Dugourd et. al., Dynamic partitioning of branched-chain amino acids-derived
nitrogen supports renal cancer progression , Nature Communications 2022, \doi{10.1038/s41467-022-35036-4}
}
\usage{
cellular_meta
}
\description{
Metabolomics workbench project PR001418, study ST002226 and ST002224 measured metabolites were assigned
HMDB and KEGG IDs as well as one main metabolic pathway.
}
\keyword{datasets}
