% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefactorPriorKnoweldge.R
\name{checkmatch_pk_to_data}
\alias{checkmatch_pk_to_data}
\title{Check and summarize prior_knowledge-to-MeasuredFeatures relationship}
\usage{
checkmatch_pk_to_data(
  data,
  input_pk,
  metadata_info = c(InputID = "HMDB", PriorID = "HMDB", grouping_variable = "term"),
  save_table = "csv",
  path = NULL
)
}
\arguments{
\item{data}{dataframe with at least one column with the detected metabolite IDs (e.g. HMDB). If there are multiple IDs per detected peak, please separate them by comma ("," or ", " or chr list). If there is a main ID and additional IDs, please provide them in separate columns.}

\item{input_pk}{dataframe with at least one column with the metabolite ID (e.g. HMDB) that need to match data metabolite IDs "source" (e.g. term). If there are multiple IDs, as the original pathway IDs (e.g. KEGG) where translated (e.g. to HMDB), please separate them by comma ("," or ", " or chr list).}

\item{metadata_info}{Colum name of Metabolite IDs in data and input_pk as well as column name of grouping_variable in input_pk. \strong{Default = c(InputID="HMDB", PriorID="HMDB", grouping_variable="term")}}

\item{save_table}{\emph{Optional: } File types for the analysis results are: "csv", "xlsx", "txt". \strong{Default = "csv"}}

\item{path}{{Optional:} Path to the folder the results should be saved at. \strong{Default = NULL}}
}
\description{
Check and summarize prior_knowledge-to-MeasuredFeatures relationship
}
\examples{
DetectedIDs <-  MetaProViz::ToyData(data="Cells_Metadata")\%>\% rownames_to_column("Metabolite") \%>\%dplyr::select("Metabolite", "HMDB")\%>\%tidyr::drop_na()
input_pathway <- MetaProViz::translate_id(data= MetaProViz::metsigdb_kegg(), metadata_info = c(InputID="MetaboliteID", grouping_variable="term"), from = c("kegg"), to = c("hmdb"))[["TranslatedDF"]]\%>\%tidyr::drop_na()
Res <- MetaProViz:::checkmatch_pk_to_data(data= DetectedIDs, input_pk= input_pathway, metadata_info = c(InputID="HMDB", PriorID="hmdb", grouping_variable="term"))

}
