% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing.R
\name{core_norm}
\alias{core_norm}
\title{Normalize consumption release data}
\usage{
core_norm(data, metadata_sample, metadata_info)
}
\arguments{
\item{data}{DF which contains unique sample identifiers as row names and metabolite
numerical values in columns with metabolite identifiers as column names.
Use NA for metabolites that were not detected and consider converting
any zeros to NA unless they are true zeros.}

\item{metadata_sample}{DF which contains information about the samples, which will be combined
with the input data based on the unique sample identifiers used as
rownames.}

\item{metadata_info}{Named vector containing the information about the names of the
experimental parameters. c(Conditions="ColumnName_Plot_SettingsFile",
core_norm_factor = "Columnname_Input_SettingsFile", core_media =
"Columnname_Input_SettingsFile"). Column core_norm_factor is used for
normalization and core_media is used to specify the name of the media
controls in the Conditions.}
}
\value{
List with two elements: DF (including output table) and Plot (including
all plots generated)

This function is intended for standalone CoRe-style experiments with
control-media metadata available via \code{metadata_info}.
}
\description{
Normalize consumption release data
}
\examples{
Media <-
    medium_raw \%>\%
    tibble::column_to_rownames("Code") \%>\%
    subset(!Conditions == "Pool") \%>\%
    dplyr::mutate_all(~ ifelse(grepl("^0*(\\\\.0*)?$", as.character(.)), NA, .))
Res <- core_norm(
    data = Media[, -c(1:3)],
    metadata_sample = Media[, c(1:3)],
    metadata_info = c(
        Conditions = "Conditions",
        core_norm_factor = "GrowthFactor",
        core_media = "blank"
    )
)

}
