% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverRepresentationAnalysis.R
\name{DM_ORA}
\alias{DM_ORA}
\title{DM_ORA}
\usage{
DM_ORA(
  Input_data,
  PathwayFile,
  PathwayName = "",
  minGSSize = 10,
  maxGSSize = 1000,
  Save_as = "svg",
  pCutoff = 0.2,
  PercentageCutoff = 10
)
}
\arguments{
\item{Input_data}{Input DF with column "t.val" and column "Metabolite", which needs to match the identifier type in column "Metabolite" in the PathwayFile.}

\item{PathwayFile}{DF that must include column "term" with the pathway name, column "Metabolite" with the Metabolite name or ID and column "Description" with pathway description that will be depicted on the plots.}

\item{PathwayName}{\emph{Optional: } Name of the pathway list used \strong{default: ""}}

\item{minGSSize}{\emph{Optional: } minimum group size in ORA \strong{default: 10}}

\item{maxGSSize}{\emph{Optional: } maximum group size in ORA \strong{default: 1000}}

\item{Save_as}{\emph{Optional: } If Save_as=NULL no plots will be saved. Otherwise, file types for the figures are: "svg", "pdf", "png" \strong{default: "pdf"}}

\item{pCutoff}{\emph{Optional: } p-adjusted value cutoff from ORA results that should be plotted. If Save_as=NULL, this is ignored. \strong{default: 0.2}}

\item{PercentageCutoff}{\emph{Optional: } Percentage cutoff of metabolites that are detected of a pathway from ORA results that should be plotted. If Save_as=NULL, this is ignored. \strong{default: 10}}
}
\value{
Saves results as individual .csv files.
}
\description{
Uses enricher to run ORA on the differential metabolites (DM) using a pathway list
}
