% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialMetaboliteAnalysis.R
\name{DMA}
\alias{DMA}
\title{This script allows you to perform differential metabolite analysis to obtain a Log2FC, pval, padj and tval comparing two conditions.}
\usage{
DMA(
  Input_data,
  Experimental_design,
  Condition1,
  Condition2,
  STAT_pval = "t.test",
  STAT_padj = "fdr",
  Input_Pathways = NULL,
  OutputName = "",
  CoRe = FALSE,
  Plot = TRUE,
  Save_as = "svg"
)
}
\arguments{
\item{Input_data}{DF with unique sample identifiers as row names and metabolite numerical values in columns with metabolite identifiers as column names. Use NA for metabolites that were not detected. includes experimental Experimental_design and outlier column.}

\item{Experimental_design}{DF which contains information about the samples, which will be combined with your input data based on the unique sample identifiers used as rownames. Column "Conditions" with information about the sample conditions (e.g. "N" and "T" or "Normal" and "Tumor"), can be used for feature filtering and colour coding in the PCA. Column "AnalyticalReplicate" including numerical values, defines technical repetitions of measurements, which will be summarised. Column "BiologicalReplicates" including numerical values. Please use the following names: "Conditions", "Biological_Replicates", "Analytical_Replicates".}

\item{STAT_pval}{\emph{Optional: } String which contains an abbreviation of the selected test to calculate p.value (t.test or wilcox.test) \strong{"t-test"}}

\item{STAT_padj}{\emph{Optional: } String which contains an abbreviation of the selected p.adjusted test for p.value correction for multiple Hypothesis testing. Search: ?p.adjust for more methods:"BH", "fdr", "bonferroni", "holm", etc.\strong{"fdr"}}

\item{Input_Pathways}{\emph{Optional: } DF which contains the pathway information for each metabolite. \strong{NULL}}

\item{OutputName}{String which is added to the output files of the DMA.}

\item{CoRe}{\emph{Optional: } TRUE or FALSE for whether a Consumption/Release  input is used \strong{FALSE}}

\item{Save_as}{\emph{Optional: } Select the file type of output plots. Options are svg, png, pdf, jpeg, tiff, bmp. \strong{Default = svg}}

\item{Conditon1}{Input needs to contain a column named "Condition" including the Condition1 that will be compared to Condition2, e.g. "KO".}

\item{Conditon2}{Input needs to contain a column named "Condition" including Condition2 that is compared to Condition1, e.g. "WT".}

\item{plot}{\emph{Optional: } TRUE or FALSE, if TRUE Volcano plot is saved as an overview of the results. \strong{TRUE}}
}
\description{
This script allows you to perform differential metabolite analysis to obtain a Log2FC, pval, padj and tval comparing two conditions.
}
\keyword{Analysis,}
\keyword{Differential}
\keyword{Hypothesis}
\keyword{Metabolite}
\keyword{Multiple}
\keyword{Normality}
\keyword{testing}
\keyword{testing,}
