% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialMetaboliteAnalysis.R
\name{dma}
\alias{dma}
\title{This function allows you to perform differential metabolite analysis to
obtain a Log2FC, pval, padj and tval comparing two or multiple conditions.}
\usage{
dma(
  data,
  metadata_sample = NULL,
  metadata_info = c(Conditions = "Conditions", Numerator = NULL, Denominator = NULL),
  pval = "lmFit",
  padj = "fdr",
  metadata_feature = NULL,
  core = FALSE,
  vst = FALSE,
  shapiro = TRUE,
  bartlett = TRUE,
  transform = TRUE,
  save_plot = "svg",
  save_table = "csv",
  print_plot = TRUE,
  path = NULL
)
}
\arguments{
\item{data}{SummarizedExperiment (se) file including assay and colData.
If se file is provided, metadata_sample is extracted from the colData
of the se object. metadata_feature, if available, are extracted from
the rowData. Alternatively provide a DF with unique sample identifiers
as row names and metabolite numerical values in columns with
metabolite identifiers as column names. Use NA for metabolites that
were not detected.}

\item{metadata_sample}{\emph{Optional: } Only required if you did not
provide se file in parameter data. Provide DF which contains metadata
information about the samples, which will be combined with your input
data based on the unique sample identifiers used as rownames.
\strong{Default = NULL}}

\item{metadata_info}{\emph{Optional: } Named vector including the
information about the conditions column information on numerator or
denominator c(Conditions="ColumnName_SettingsFile", Numerator =
"ColumnName_SettingsFile", Denominator  = "ColumnName_SettingsFile").
Denominator and Numerator will specify which comparison(s) will be
done (one-vs-one, all-vs-one, all-vs-all), e.g. Denominator=NULL and
Numerator =NULL selects all the condition and performs multiple
comparison all-vs-all. Log2FC are obtained by dividing the numerator
by the denominator, thus positive Log2FC values mean higher expression
in the numerator. \strong{Default = c(conditions="Conditions",
numerator = NULL, denumerator = NULL)}}

\item{pval}{\emph{Optional: } String which contains an abbreviation of the
selected test to calculate p.value. For one-vs-one comparisons choose
t.test, wilcox.test, "chisq.test", "cor.test" or lmFit (=limma), for
one-vs-all or all-vs-all comparison choose aov (=anova), welch(=welch
anova), kruskal.test or lmFit (=limma) \strong{Default = "lmFit"}}

\item{padj}{\emph{Optional: } String which contains an abbreviation of the
selected p.adjusted test for p.value correction for multiple
Hypothesis testing. Search: ?p.adjust for more methods:"BH", "fdr",
"bonferroni", "holm", etc.\strong{Default = "fdr"}}

\item{metadata_feature}{\emph{Optional: } To provide metadata information ,
i.e. pathway information, retention time,..., for each metabolite.
Only used if you did not provide se file in parameter data. Provide
DF where the row names must match the metabolite names in the
columns of the data. \strong{Default = NULL}}

\item{core}{\emph{Optional: } TRUE or FALSE for whether a Consumption/Release
input is used. \strong{Default = FALSE}}

\item{vst}{TRUE or FALSE for whether to use variance stabilizing
transformation on the data when linear modeling is used for hypothesis
testing. \strong{Default = FALSE}}

\item{shapiro}{TRUE or FALSE for whether to perform the shapiro.test and get
informed about data distribution (normal versus not-normal
distribution. \strong{Default = TRUE}}

\item{bartlett}{TRUE or FALSE for whether to perform the bartlett.test.
\strong{Default = TRUE}}

\item{transform}{TRUE or FALSE. If TRUE we expect the data to be not log2
transformed and log2 transformation will be performed within the limma
function and Log2FC calculation. If FALSE we expect the data to be
log2 transformed as this impacts the Log2FC calculation and limma.
\strong{Default= TRUE}}

\item{save_plot}{\emph{Optional: } Select the file type of output plots.
Options are svg, png, pdf. \strong{Default = svg}}

\item{save_table}{\emph{Optional: } File types for the analysis results are:
"csv", "xlsx", "txt". \strong{Default = "csv"}}

\item{print_plot}{\emph{Optional: } TRUE or FALSE, if TRUE Volcano plot is
saved as an overview of the results. \strong{Default = TRUE}}

\item{path}{\emph{Optional:} Path to the folder the results should be saved
at. \strong{Default = NULL}}
}
\value{
Dependent on parameter settings, list of lists will be returned for
dma (DF of each comparison), shapiro (Includes DF and Plot), bartlett
(Includes DF and Histogram), vst (Includes DF and Plot) and
VolcanoPlot (Plots of each comparison).
}
\description{
This function allows you to perform differential metabolite analysis to
obtain a Log2FC, pval, padj and tval comparing two or multiple conditions.
}
\examples{
data(intracell_raw_se)
ResI <- dma(
    data = intracell_raw_se,
    metadata_info = c(
        Conditions = "Conditions", Numerator = NULL, Denominator = "HK2"
    )
)

data(intracell_raw)
Intra <- intracell_raw[-c(49:58), ] \%>\% tibble::column_to_rownames("Code")
ResI <- dma(
    data = Intra[, -c(1:3)],
    metadata_sample = Intra[, c(1:3)],
    metadata_info = c(
        Conditions = "Conditions", Numerator = NULL, Denominator = "HK2"
    )
)

}
\keyword{Analysis,}
\keyword{Differential}
\keyword{Hypothesis}
\keyword{Metabolite}
\keyword{Multiple}
\keyword{Normality}
\keyword{testing}
\keyword{testing,}
