% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyData.R
\docType{data}
\name{equivalent_features}
\alias{equivalent_features}
\title{equivalent_features}
\format{
Columns include metabolite trivial names, metabolite IDs (HMDB, KEGG, etc.), metabolite structural information (=INCHI).
}
\source{
Schmidt et al, MetaProViz: METabolomics pre-PRocessing, functiOnal analysis and VIZualisation version 2.1.7, GitHub 2025.
}
\usage{
equivalent_features
}
\description{
Manually curated list of aminoacids and aminoacid-related metabolites with corresponding metabolite
identifiers (HMDB, KEGG, etc.) irrespective of chirality.
}
\examples{
data(equivalent_features)
head(equivalent_features)
}
\keyword{datasets}
