% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefactorPriorKnoweldge.R
\name{equivalent_id}
\alias{equivalent_id}
\title{Find additional potential IDs for  "kegg", "pubchem", "chebi", "hmdb"}
\usage{
equivalent_id(
  data,
  metadata_info = c(InputID = "MetaboliteID"),
  from = "hmdb",
  save_table = "csv",
  path = NULL
)
}
\arguments{
\item{data}{dataframe with at least one column with the detected metabolite IDs (one ID per row).}

\item{metadata_info}{\emph{Optional: } Column name of metabolite IDs. \strong{Default = list(InputID="MetaboliteID")}}

\item{from}{ID type that is present in your data. Choose between "kegg", "pubchem", "chebi", "hmdb". \strong{Default = "hmdb"}}

\item{save_table}{\emph{Optional: } File types for the analysis results are: "csv", "xlsx", "txt". \strong{Default = "csv"}}

\item{path}{{Optional:} Path to the folder the results should be saved at. \strong{Default = NULL}}
}
\value{
Input DF with additional column including potential additional IDs.
}
\description{
Find additional potential IDs for  "kegg", "pubchem", "chebi", "hmdb"
}
\examples{
DetectedIDs <- cellular_meta\%>\%tidyr::drop_na()
Res <- MetaProViz::equivalent_id(data= DetectedIDs, metadata_info = c(InputID="HMDB"), from = "hmdb")

}
\keyword{Find}
\keyword{IDs}
\keyword{additional}
\keyword{for}
\keyword{identifier}
\keyword{metabolite}
\keyword{one}
\keyword{potential}
