% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing.R
\name{feature_filtering}
\alias{feature_filtering}
\title{Filter features}
\usage{
feature_filtering(
  data,
  metadata_sample,
  metadata_info,
  core = FALSE,
  featurefilt = "Modified",
  cutoff_featurefilt = 0.8
)
}
\arguments{
\item{data}{DF which contains unique sample identifiers as row names and metabolite
numerical values in columns with metabolite identifiers as column names.
Use NA for metabolites that were not detected and consider converting
any zeros to NA unless they are true zeros.}

\item{metadata_sample}{DF which contains information about the samples, which will be combined
with the input data based on the unique sample identifiers used as
rownames.}

\item{metadata_info}{Named vector containing the information about the names of the
experimental parameters. c(Conditions="ColumnName_Plot_SettingsFile",
Biological_Replicates="ColumnName_Plot_SettingsFile"). Column
"Conditions" with information about the sample conditions (e.g. "N" and
"T" or "Normal" and "Tumor"), can be used for feature filtering and
colour coding in the PCA. Column "BiologicalReplicates" including
numerical values. For core = TRUE add core_media =
"Columnname_Input_SettingsFile", which specifies the name of the media
controls in the Conditions.}

\item{core}{\emph{Optional: } If TRUE, a consumption-release experiment has been
performed.Should not be normalised to media blank. Provide information
about control media sample names via metadata_info "core_media" samples.
\strong{Default = FALSE}}

\item{featurefilt}{\emph{Optional: } If NULL, no feature filtering is performed. If set to
"Standard" then it applies the 80\%-filtering rule (Bijlsma S. et al.,
2006) on the metabolite features on the whole dataset. If is set to
"Modified",filtering is done based on the different conditions, thus a
column named "Conditions" must be provided in the Input_SettingsFile
input file including the individual conditions you want to apply the
filtering to (Yang, J et al., 2015). \strong{Default = Modified}}

\item{cutoff_featurefilt}{\emph{Optional: } percentage of feature filtering. \strong{Default =
0.8}}
}
\value{
List with two elements: filtered matrix  and features filtered

This function can be used as a standalone preprocessing step on raw input
data with matching sample metadata.
}
\description{
Filter features
}
\examples{
data(intracell_raw)
Intra <- intracell_raw \%>\% tibble::column_to_rownames("Code")
Res <- feature_filtering(
    data =
        Intra[-c(49:58), -c(1:3)] \%>\%
        dplyr::mutate_all(
            ~ ifelse(grepl("^0*(\\\\.0*)?$", as.character(.)), NA, .)
        ),
    metadata_sample = Intra[-c(49:58), c(1:3)],
    metadata_info = c(
        Conditions = "Conditions",
        Biological_Replicates = "Biological_Replicates"
    )
)

}
