% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyData.R
\docType{data}
\name{gaude_pathways}
\alias{gaude_pathways}
\title{gaude_pathways}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1932 rows and 3 columns.
}
\usage{
gaude_pathways
}
\description{
gaude_pathways
}
\examples{
data(gaude_pathways)
head(gaude_pathways)

}
\keyword{datasets}
