% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyData.R
\docType{data}
\name{intracell_dma}
\alias{intracell_dma}
\title{intracell_dma}
\format{
Columns include Log2FC, stats, metabolite identifiers, metabolite pathways and normalised
metabolite values used as input with row names being metabolitetrivial names.
}
\source{
Sciacovelli & Dugourd et. al., Dynamic partitioning of branched-chain amino acids-derived
nitrogen supports renal cancer progression , Nature Communications 2022, \doi{10.1038/s41467-022-35036-4}
}
\usage{
intracell_dma
}
\description{
Metabolomics workbench project PR001418, study ST002224 where we performed differential metabolite
analysis comparing intracellular metabolomics of 786-M1A versus HK2 cells.
}
\keyword{datasets}
