% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyData.R
\docType{data}
\name{intracell_raw}
\alias{intracell_raw}
\title{intracell_raw}
\format{
A data frame with multiple rows and columns:
\describe{
\item{Conditions}{Character vector indicating cell line identity}
\item{Analytical_Replicate}{Integer replicate number for analytical replicates}
\item{Biological_Replicate}{Integer replicate number for biological replicates}
\item{...}{Numeric columns for each measured metabolite (raw peak values)}
}
}
\source{
Sciacovelli & Dugourd et. al., Dynamic partitioning of branched-chain amino acids-derived
nitrogen supports renal cancer progression, Nature Communications 2022, DOI:10.1038/s41467-022-35036-4.
}
\usage{
intracell_raw
}
\description{
Metabolomics workbench project PR001418, study ST002224 where we exported integrated raw
peak values of intracellular metabolomics of HK2 and ccRCC cell lines 786-O, 786-M1A and 786-M2A.
}
\examples{
data(intracell_raw)
head(intracell_raw)
}
\keyword{datasets}
