% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPriorKnoweldge.R
\name{LoadHallmarks}
\alias{LoadHallmarks}
\title{Toy Data Import}
\usage{
LoadHallmarks()
}
\value{
A data frame containing the toy data.
}
\description{
Import and process .csv file to create toy data.
}
