% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPriorKnoweldge.R
\name{Make_GeneMetabSet}
\alias{Make_GeneMetabSet}
\title{Function to add metabolite HMDB IDs to existing genesets based on cosmosR prior knowledge}
\usage{
Make_GeneMetabSet(
  Input_GeneSet,
  SettingsInfo = c(Target = "gene"),
  PKName = NULL,
  SaveAs_Table = "csv",
  FolderPath = NULL
)
}
\arguments{
\item{Input_GeneSet}{Dataframe with two columns for source (=term) and Target (=gene), e.g. Hallmarks.}

\item{SettingsInfo}{\emph{Optional: }  Column name of Target in Input_GeneSet. \strong{Default = c(Target="gene")}}

\item{PKName}{\emph{Optional: } Name of the prior knowledge resource. \strong{default: NULL}}

\item{SaveAs_Table}{\emph{Optional: } File types for the analysis results are: "csv", "xlsx", "txt". \strong{Default = "csv"}}

\item{FolderPath}{{Optional:} String which is added to the resulting folder name \strong{default: NULL}}
}
\description{
Function to add metabolite HMDB IDs to existing genesets based on cosmosR prior knowledge
}
