% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPriorKnoweldge.R
\name{Make_GeneMetabSet}
\alias{Make_GeneMetabSet}
\title{Function to add metabolite HMDB IDs to existing genesets based on cosmosR prior knowledge}
\usage{
Make_GeneMetabSet(
  Input_GeneSet,
  target = c("gene"),
  OutputName = NULL,
  Folder_Name = NULL
)
}
\arguments{
\item{Input_GeneSet}{Dataframe with two columns for source (=term) and target (=gene), e.g. Hallmarks.}

\item{target}{\emph{Optional: }  Column name of target in Input_GeneSet. \strong{Default = "gene}}

\item{OutputName}{{Optional:} String which is added to the output files name.\strong(Default = NULL)}

\item{Folder_Name}{{Optional:} String which is added to the resulting folder name \strong(Default = NULL)}
}
\description{
Function to add metabolite HMDB IDs to existing genesets based on cosmosR prior knowledge
}
