% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboliteClusteringAnalysis.R
\name{MCA_2Cond}
\alias{MCA_2Cond}
\title{\subsection{---------------------------}{

MCA_2Cond
}}
\usage{
MCA_2Cond(
  Cond1_File,
  Cond2_File,
  MetaboliteID = "Metabolite",
  Cond1ValueCol = "Log2FC",
  Cond1PadjCol = "p.adj",
  Cond2ValueCol = "Log2FC",
  Cond2PadjCol = "p.adj",
  Cond1_padj_cutoff = 0.05,
  Cond2_padj_cutoff = 0.05,
  Cond1_FC_cutoff = 1,
  Cond2_FC_cutoff = 1,
  Save_as_Results = "xlsx",
  backgroundMethod = "C1&C2",
  OutputFileName = "MCA_2Cond_"
)
}
\arguments{
\item{Cond1_File}{DF for your data (results from e.g. DMA) containing metabolites in rows with corresponding Log2FC and stat (p-value, p.adjusted) value columns.}

\item{Cond2_File}{DF for your data (results from e.g. DMA) containing metabolites in rows with corresponding Log2FC and stat (p-value, p.adjusted) value columns.}

\item{MetaboliteID}{Column name of Column including the Metabolite identifiers. This MUST BE THE SAME in each of your Input files.}

\item{Cond1ValueCol}{Column name of Log2FC in Cond1File}

\item{Cond1PadjCol}{Column name of adjusted p-value in Cond1File. Can also be you p-value column if you want to use this instead.}

\item{Cond2ValueCol}{Column name of Log2FC in Cond2File}

\item{Cond2PadjCol}{Column name of adjusted p-value in Cond2File. Can also be you p-value column if you want to use this instead.}

\item{Cond1_padj_cutoff}{\emph{Optional: } adjusted p-value cutoff for Cond1File. \strong{Default=0.05}}

\item{Cond2_padj_cutoff}{\emph{Optional: } adjusted p-value cutoff for Cond2File. \strong{Default=0.05}}

\item{Cond1_FC_cutoff}{\emph{Optional: } Log2FC cutoff for Cond1File. \strong{Default=0.5}}

\item{Cond2_FC_cutoff}{\emph{Optional: } Log2FC cutoff for Cond2File. \strong{Default=0.5}}

\item{Save_as_Results}{\emph{Optional: } File types for the analysis results are: "csv", "xlsx", "txt" \strong{default: "xlsx"}}

\item{backgroundMethod}{\emph{Optional: } Background method C1|C2, C1&C2, C2, C1 or * \strong{Default="C1&C2"}}

\item{outputFileName}{\emph{Optional: } Output filename \strong{Default=SiRCle_RCM.csv}}
}
\value{
MCA an instance of the MetaProViz package
}
\description{
This script allows you to perform metabolite clustering analysis and computes clusters of metabolites based on regulatory rules between two conditions (e.g. KO versus WT in Hypoxia = Cond1 and KO versus WT in Normoxia = Cond2).
}
