% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboliteClusteringAnalysis.R
\name{MCA_2Cond}
\alias{MCA_2Cond}
\title{\subsection{---------------------------}{

This script performs metabolite clustering analysis and computes clusters of metabolites based on regulatory rules between conditions.
}}
\usage{
MCA_2Cond(
  InputData_C1,
  InputData_C2,
  SettingsInfo_C1 = c(ValueCol = "Log2FC", StatCol = "p.adj", StatCutoff = 0.05,
    ValueCutoff = 1),
  SettingsInfo_C2 = c(ValueCol = "Log2FC", StatCol = "p.adj", StatCutoff = 0.05,
    ValueCutoff = 1),
  FeatureID = "Metabolite",
  SaveAs_Table = "csv",
  BackgroundMethod = "C1&C2",
  FolderPath = NULL
)
}
\arguments{
\item{InputData_C1}{DF for your data (results from e.g. DMA) containing metabolites in rows with corresponding Log2FC and stat (p-value, p.adjusted) value columns.}

\item{InputData_C2}{DF for your data (results from e.g. DMA) containing metabolites in rows with corresponding Log2FC and stat (p-value, p.adjusted) value columns.}

\item{SettingsInfo_C1}{\emph{Optional: } Pass ColumnNames and Cutoffs for condition 1 including the value column (e.g. Log2FC, Log2Diff, t.val, etc) and the stats column (e.g. p.adj, p.val). This must include: c(ValueCol=ColumnName_InputData_C1,StatCol=ColumnName_InputData_C1, StatCutoff= NumericValue, ValueCutoff=NumericValue) \strong{Default=c(ValueCol="Log2FC",StatCol="p.adj", StatCutoff= 0.05, ValueCutoff=1)}}

\item{SettingsInfo_C2}{\emph{Optional: } Pass ColumnNames and Cutoffs for condition 2 includingthe value column (e.g. Log2FC, Log2Diff, t.val, etc) and the stats column (e.g. p.adj, p.val). This must include: c(ValueCol=ColumnName_InputData_C2,StatCol=ColumnName_InputData_C2, StatCutoff= NumericValue, ValueCutoff=NumericValue)\strong{Default=c(ValueCol="Log2FC",StatCol="p.adj", StatCutoff= 0.05, ValueCutoff=1)}}

\item{FeatureID}{\emph{Optional: } Column name of Column including the Metabolite identifiers. This MUST BE THE SAME in each of your Input files. \strong{Default="Metabolite"}}

\item{SaveAs_Table}{\emph{Optional: } File types for the analysis results are: "csv", "xlsx", "txt" \strong{Default = "csv"}}

\item{BackgroundMethod}{\emph{Optional: } Background method C1|C2, C1&C2, C2, C1 or * \strong{Default="C1&C2"}}

\item{FolderPath}{\emph{Optional:} Path to the folder the results should be saved at. \strong{Default = NULL}}
}
\value{
List of two DFs: 1. Summary of the cluster count and 2. the detailed information of each metabolites in the clusters.
}
\description{
\subsection{---------------------------}{

This script performs metabolite clustering analysis and computes clusters of metabolites based on regulatory rules between conditions.
}
}
\examples{
Intra <- MetaProViz::ToyData("IntraCells_Raw")
Input <- MetaProViz::DMA(InputData=Intra[-c(49:58) ,-c(1:3)], SettingsFile_Sample=Intra[-c(49:58) , c(1:3)], SettingsInfo = c(Conditions = "Conditions", Numerator = NULL, Denominator  = "HK2"))

Res <- MetaProViz::MCA_2Cond(InputData_C1 = Input[["DMA"]][["786-O_vs_HK2"]],
                             InputData_C2 = Input[["DMA"]][["786-M1A_vs_HK2"]])

}
\keyword{biological}
\keyword{clustering}
