% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboliteClusteringAnalysis.R, R/ToyData.R
\docType{data}
\name{mca_2cond}
\alias{mca_2cond}
\title{\subsection{---------------------------}{

This script performs metabolite clustering analysis and computes clusters of metabolites based on regulatory rules between conditions.
}}
\format{
Columns include Intra, core, core_Detection including state entries (e.g. up, down, etc.) and the Regulator Clustering
columns (RG1-RG3)
}
\source{
Schmidt et al, MetaProViz: METabolomics pre-PRocessing, functiOnal analysis and VIZualisation version 2.1.7, GitHub 2025.
}
\usage{
mca_2cond(
  data_C1,
  data_C2,
  metadata_info_C1 = c(ValueCol = "Log2FC", StatCol = "p.adj", cutoff_stat = 0.05,
    ValueCutoff = 1),
  metadata_info_C2 = c(ValueCol = "Log2FC", StatCol = "p.adj", cutoff_stat = 0.05,
    ValueCutoff = 1),
  feature = "Metabolite",
  save_table = "csv",
  method_background = "C1&C2",
  path = NULL
)

mca_2cond
}
\arguments{
\item{data_C1}{DF for your data (results from e.g. dma) containing metabolites in rows with corresponding Log2FC and stat (p-value, p.adjusted) value columns.}

\item{data_C2}{DF for your data (results from e.g. dma) containing metabolites in rows with corresponding Log2FC and stat (p-value, p.adjusted) value columns.}

\item{metadata_info_C1}{\emph{Optional: } Pass ColumnNames and Cutoffs for condition 1 including the value column (e.g. Log2FC, Log2Diff, t.val, etc) and the stats column (e.g. p.adj, p.val). This must include: c(ValueCol=ColumnName_data_C1,StatCol=ColumnName_data_C1, cutoff_stat= NumericValue, ValueCutoff=NumericValue) \strong{Default=c(ValueCol="Log2FC",StatCol="p.adj", cutoff_stat= 0.05, ValueCutoff=1)}}

\item{metadata_info_C2}{\emph{Optional: } Pass ColumnNames and Cutoffs for condition 2 includingthe value column (e.g. Log2FC, Log2Diff, t.val, etc) and the stats column (e.g. p.adj, p.val). This must include: c(ValueCol=ColumnName_data_C2,StatCol=ColumnName_data_C2, cutoff_stat= NumericValue, ValueCutoff=NumericValue)\strong{Default=c(ValueCol="Log2FC",StatCol="p.adj", cutoff_stat= 0.05, ValueCutoff=1)}}

\item{feature}{\emph{Optional: } Column name of Column including the Metabolite identifiers. This MUST BE THE SAME in each of your Input files. \strong{Default="Metabolite"}}

\item{save_table}{\emph{Optional: } File types for the analysis results are: "csv", "xlsx", "txt" \strong{Default = "csv"}}

\item{method_background}{\emph{Optional: } Background method C1|C2, C1&C2, C2, C1 or * \strong{Default="C1&C2"}}

\item{path}{\emph{Optional:} Path to the folder the results should be saved at. \strong{Default = NULL}}
}
\value{
List of two DFs: 1. summary of the cluster count and 2. the detailed information of each metabolites in the clusters.
}
\description{
Manually curated table defining the flow of information of the two condition biological regulatory clusters Regulatory labels
from the different grouping methods.
}
\examples{
Intra <- MetaProViz::ToyData("IntraCells_Raw")
Input <- MetaProViz::dma(data=Intra[-c(49:58) ,-c(1:3)], metadata_sample=Intra[-c(49:58) , c(1:3)], metadata_info = c(Conditions = "Conditions", Numerator = NULL, Denominator  = "HK2"))

Res <- MetaProViz::mca_2cond(data_C1 = Input[["dma"]][["786-O_vs_HK2"]],
                             data_C2 = Input[["dma"]][["786-M1A_vs_HK2"]])

}
\keyword{biological}
\keyword{clustering}
\keyword{datasets}
