% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboliteClusteringAnalysis.R
\name{MCA_CoRe}
\alias{MCA_CoRe}
\title{MCA_CoRe}
\usage{
MCA_CoRe(
  Intra_File,
  CoRe_File,
  MetaboliteID = "Metabolite",
  IntraValueCol = "Log2FC",
  IntraPadjCol = "p.adj",
  CoReValueCol = "Log2(Distance)",
  CoReDirectionCol = "CoRe",
  CoRePadjCol = "p.adj",
  Intra_padj_cutoff = 0.05,
  CoRe_padj_cutoff = 0.05,
  Intra_FC_cutoff = 1,
  Save_as_Results = "xlsx",
  CoRe_FC_cutoff = 20,
  backgroundMethod = "Intra&CoRe",
  OutputFileName = "MCA_",
  Folder_Name = NULL
)
}
\arguments{
\item{Intra_File}{DF for your data (results from e.g. DMA) containing metabolites in rows with corresponding Log2FC and stat (p-value, p.adjusted) value columns.}

\item{CoRe_File}{DF for your data (results from e.g. DMA) containing metabolites in rows with corresponding Log2FC and stat (p-value, p.adjusted) value columns. Here we additionally require}

\item{MetaboliteID}{Column name of Column including the Metabolite identifiers. This MUST BE THE SAME in each of your Input files.}

\item{IntraValueCol}{Column name of Log2FC in IntraFile}

\item{IntraPadjCol}{Column name of adjusted p-value in IntraFile. Can also be you p-value column if you want to use this instead.}

\item{CoReValueCol}{Column name of Log2(Distance) in CoReFile}

\item{CoRePadjCol}{Column name of adjusted p-value in CoReFile. Can also be you p-value column if you want to use this instead.}

\item{Intra_padj_cutoff}{\emph{Optional: } adjusted p-value cutoff for IntraFile. \strong{Default=0.05}}

\item{CoRe_padj_cutoff}{\emph{Optional: } adjusted p-value cutoff for CoReFile. \strong{Default=0.05}}

\item{Intra_FC_cutoff}{\emph{Optional: } Log2FC cutoff for IntraFile. \strong{Default=0.5}}

\item{CoRe_FC_cutoff}{\emph{Optional: } Log2FC cutoff for CoReFile. \strong{Default=20}}

\item{backgroundMethod}{\emph{Optional: } Background method `Intra|CoRe, Intra&CoRe, CoRe, Intra or * \strong{Default="Intra&CoRe"}}

\item{Folder_Name}{{Optional:} String which is added to the resulting folder name \strong(Default = NULL)}

\item{outputFileName}{\emph{Optional: } Output filename \strong{Default="MCA_"}}
}
\value{
MCA an instance of the MetaProViz package
}
\description{
This script allows you to perform metabolite clustering analysis and computes clusters of metabolites based on regulatory rules between Intracellular and culture media metabolomics (CoRe experiment).
}
