% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboliteClusteringAnalysis.R
\name{MCA_CoRe}
\alias{MCA_CoRe}
\title{This script performs metabolite clustering analysis and computes clusters of metabolites based on regulatory rules between Intracellular and culture media metabolomics (CoRe experiment).}
\usage{
MCA_CoRe(
  InputData_Intra,
  InputData_CoRe,
  SettingsInfo_Intra = c(ValueCol = "Log2FC", StatCol = "p.adj", StatCutoff = 0.05,
    ValueCutoff = 1),
  SettingsInfo_CoRe = c(DirectionCol = "CoRe", ValueCol = "Log2(Distance)", StatCol =
    "p.adj", StatCutoff = 0.05, ValueCutoff = 1),
  FeatureID = "Metabolite",
  SaveAs_Table = "csv",
  BackgroundMethod = "Intra&CoRe",
  FolderPath = NULL
)
}
\arguments{
\item{InputData_Intra}{DF for your data (results from e.g. DMA) containing metabolites in rows with corresponding Log2FC and stat (p-value, p.adjusted) value columns.}

\item{InputData_CoRe}{DF for your data (results from e.g. DMA) containing metabolites in rows with corresponding Log2FC and stat (p-value, p.adjusted) value columns. Here we additionally require}

\item{SettingsInfo_Intra}{\emph{Optional: } Pass ColumnNames and Cutoffs for the intracellular metabolomics including the value column (e.g. Log2FC, Log2Diff, t.val, etc) and the stats column (e.g. p.adj, p.val). This must include: c(ValueCol=ColumnName_InputData_Intra,StatCol=ColumnName_InputData_Intra, StatCutoff= NumericValue, ValueCutoff=NumericValue) \strong{Default=c(ValueCol="Log2FC",StatCol="p.adj", StatCutoff= 0.05, ValueCutoff=1)}}

\item{SettingsInfo_CoRe}{\emph{Optional: } Pass ColumnNames and Cutoffs for the consumption-release metabolomics including the direction column, the value column (e.g. Log2Diff, t.val, etc) and the stats column (e.g. p.adj, p.val). This must include: c(DirectionCol= ColumnName_InputData_CoRe,ValueCol=ColumnName_InputData_CoRe,StatCol=ColumnName_InputData_CoRe, StatCutoff= NumericValue, ValueCutoff=NumericValue)\strong{Default=c(DirectionCol="CoRe", ValueCol="Log2(Distance)",StatCol="p.adj", StatCutoff= 0.05, ValueCutoff=1)}}

\item{FeatureID}{\emph{Optional: } Column name of Column including the Metabolite identifiers. This MUST BE THE SAME in each of your Input files. \strong{Default="Metabolite"}}

\item{SaveAs_Table}{\emph{Optional: } File types for the analysis results are: "csv", "xlsx", "txt" \strong{default: "csv"}}

\item{BackgroundMethod}{\emph{Optional: } Background method `Intra|CoRe, Intra&CoRe, CoRe, Intra or * \strong{Default="Intra&CoRe"}}

\item{FolderPath}{\emph{Optional:} Path to the folder the results should be saved at. \strong{default: NULL}}
}
\value{
List of two DFs: 1. Summary of the cluster count and 2. the detailed information of each metabolites in the clusters.
}
\description{
This script performs metabolite clustering analysis and computes clusters of metabolites based on regulatory rules between Intracellular and culture media metabolomics (CoRe experiment).
}
\examples{

Media <- ToyData("CultureMedia_Raw")
ResM <- PreProcessing(InputData = Media[-c(40:45) ,-c(1:3)],
                                  SettingsFile_Sample = Media[-c(40:45) ,c(1:3)] ,
                                  SettingsInfo = c(Conditions = "Conditions", Biological_Replicates = "Biological_Replicates", CoRe_norm_factor = "GrowthFactor", CoRe_media = "blank"),
                                  CoRe=TRUE)

MediaDMA <- DMA(InputData=ResM[["DF"]][["Preprocessing_output"]][ ,-c(1:4)],
                            SettingsFile_Sample=ResM[["DF"]][["Preprocessing_output"]][ , c(1:4)],
                            SettingsInfo = c(Conditions = "Conditions", Numerator = NULL, Denominator  = "HK2"),
                            StatPval ="aov",
                            CoRe=TRUE)

IntraDMA <- ToyData(Data="IntraCells_DMA")

Res <- MCA_CoRe(InputData_Intra = IntraDMA\%>\%tibble::rownames_to_column("Metabolite"),
                            InputData_CoRe = MediaDMA[["DMA"]][["786-M1A_vs_HK2"]])

}
\keyword{biological}
\keyword{clustering}
