% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboliteClusteringAnalysis.R
\name{mca_core}
\alias{mca_core}
\title{Metabolite clustering analysis for core experiments}
\usage{
mca_core(
  data_intra,
  data_core,
  metadata_info_intra = c(ValueCol = "Log2FC", StatCol = "p.adj", cutoff_stat = 0.05,
    ValueCutoff = 1),
  metadata_info_core = c(DirectionCol = "core", ValueCol = "Log2(Distance)", StatCol =
    "p.adj", cutoff_stat = 0.05, ValueCutoff = 1),
  feature = "Metabolite",
  save_table = "csv",
  method_background = "Intra&core",
  path = NULL
)
}
\arguments{
\item{data_intra}{DF for your data (results from e.g. dma) containing metabolites in rows
with corresponding Log2FC and stat (p-value, p.adjusted) value columns.}

\item{data_core}{DF for your data (results from e.g. dma) containing metabolites in rows
with corresponding Log2FC and stat (p-value, p.adjusted) value columns.
Here we additionally require}

\item{metadata_info_intra}{\emph{Optional: } Pass ColumnNames and Cutoffs for the intracellular
metabolomics including the value column (e.g. Log2FC, Log2Diff, t.val,
etc) and the stats column (e.g. p.adj, p.val). This must include:
c(ValueCol=ColumnName_data_intra,StatCol=ColumnName_data_intra,
cutoff_stat= NumericValue, ValueCutoff=NumericValue)
\strong{Default=c(ValueCol="Log2FC",StatCol="p.adj", cutoff_stat= 0.05,
ValueCutoff=1)}}

\item{metadata_info_core}{\emph{Optional: } Pass ColumnNames and Cutoffs for the
consumption-release metabolomics including the direction column, the
value column (e.g. Log2Diff, t.val, etc) and the stats column (e.g.
p.adj, p.val). This must include: c(DirectionCol=
ColumnName_data_core,ValueCol=ColumnName_data_core,StatCol=ColumnName_data_core,
cutoff_stat= NumericValue,
ValueCutoff=NumericValue)\strong{Default=c(DirectionCol="core",
ValueCol="Log2(Distance)",StatCol="p.adj", cutoff_stat= 0.05,
ValueCutoff=1)}}

\item{feature}{\emph{Optional: } Column name of Column including the Metabolite
identifiers. This MUST BE THE SAME in each of your Input files.
\strong{Default="Metabolite"}}

\item{save_table}{\emph{Optional: } File types for the analysis results are: "csv",
"xlsx", "txt" \strong{default: "csv"}}

\item{method_background}{\emph{Optional: } Background method `Intra|core, Intra&core, core, Intra
or * \strong{Default="Intra&core"}}

\item{path}{\emph{Optional:} Path to the folder the results should be saved at.
\strong{default: NULL}}
}
\value{
List of two DFs: 1. summary of the cluster count and 2. the detailed
information of each metabolites in the clusters.
}
\description{
Performs metabolite clustering analysis and computes clusters based on
regulatory rules between intracellular and culture media metabolomics in
core experiments.
}
\examples{
data(medium_raw)
Media <- medium_raw \%>\% tibble::column_to_rownames("Code")
ResM <- processing(
    data = Media[-c(40:45), -c(1:3)],
    metadata_sample = Media[-c(40:45), c(1:3)],
    metadata_info = c(
        Conditions = "Conditions",
        Biological_Replicates = "Biological_Replicates",
        core_norm_factor = "GrowthFactor",
        core_media = "blank"
    ),
    core = TRUE
)

MediaDMA <- dma(
    data = ResM[["DF"]][["Preprocessing_output"]][, -c(1:4)],
    metadata_sample = ResM[["DF"]][["Preprocessing_output"]][, c(1:4)],
    metadata_info = c(
        Conditions = "Conditions",
        Numerator = NULL,
        Denominator = "HK2"
    ),
    pval = "aov",
    core = TRUE
)

data(intracell_dma)
IntraDMA <- intracell_dma

Res <- mca_core(
    data_intra = as.data.frame(IntraDMA),
    data_core = as.data.frame(MediaDMA[["dma"]][["786-M1A_vs_HK2"]])
)

}
