% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyData.R
\docType{data}
\name{mca_core_rules}
\alias{mca_core_rules}
\title{mca_core_rules}
\format{
Columns include Intra, core, core_Detection including state entries (e.g. up, down, etc.) and the Regulator Clustering
columns (RG1-RG3)
}
\source{
Schmidt et al, MetaProViz: METabolomics pre-PRocessing, functiOnal analysis and VIZualisation version 2.1.7, GitHub 2025.
}
\usage{
mca_core_rules
}
\description{
Manually curated table defining the flow of information of the Conusuption-Release and Intracellular metabolomics biological
regulatory clusters Regulatory labels from the different grouping methods.
}
\examples{
data(mca_core_rules)
head(mca_core_rules)
}
\keyword{datasets}
