% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboliteClusteringAnalysis.R
\name{MCA_rules}
\alias{MCA_rules}
\title{MCA regulatory rules Import}
\usage{
MCA_rules(Method)
}
\arguments{
\item{Method}{Either "2Cond" or "CoRe" depending which regulatory rules you would like to load}
}
\value{
A data frame containing the toy data.
}
\description{
Import and process .csv file to create toy data.
}
\details{
Imports MCA regulatory rules into environment
}
