% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaDataAnalysis.R
\name{MetaAnalysis}
\alias{MetaAnalysis}
\title{This script allows you to perform}
\usage{
MetaAnalysis(
  InputData,
  SettingsFile_Sample,
  Scaling = TRUE,
  Percentage = 0.1,
  StatCutoff = 0.05,
  SaveAs_Table = "csv",
  SaveAs_Plot = "svg",
  PrintPlot = TRUE,
  FolderPath = NULL
)
}
\arguments{
\item{InputData}{DF with unique sample identifiers as row names and metabolite numerical values in columns with metabolite identifiers as column names. Use NA for metabolites that were not detected. includes experimental design and outlier column.}

\item{SettingsFile_Sample}{\emph{Optional: } DF which contains information about the samples, which will be combined with your input data based on the unique sample identifiers used as rownames. Column "Conditions" with information about the sample conditions (e.g. "N" and "T" or "Normal" and "Tumor"), can be used for feature filtering and colour coding in the PCA. Column "AnalyticalReplicate" including numerical values, defines technical repetitions of measurements, which will be summarised. Column "BiologicalReplicates" including numerical values. Please use the following names: "Conditions", "Biological_Replicates", "Analytical_Replicates".\strong{Default = NULL}}

\item{Scaling}{\emph{Optional: } TRUE or FALSE for whether a data scaling is used \strong{Default = TRUE}}

\item{Percentage}{\emph{Optional: } Percentage of top and bottom features to be displayed in the results. \strong{Default = 0.1}}

\item{StatCutoff}{\emph{Optional: } Cutoff for the adjusted p-value of the ANOVA test. \strong{Default = 0.05}}

\item{SaveAs_Table}{\emph{Optional: } File types for the analysis results are: "csv", "xlsx", "txt". \strong{Default = "csv"}}

\item{SaveAs_Plot}{\emph{Optional: } Select the file type of output plots. Options are svg, png, pdf. \strong{Default = svg}}

\item{PrintPlot}{\emph{Optional: } TRUE or FALSE, if TRUE Volcano plot is saved as an overview of the results. \strong{Default = TRUE}}

\item{FolderPath}{\emph{Optional:} Path to the folder the results should be saved at. \strong{default: NULL}}
}
\description{
This script allows you to perform
}
\keyword{PCA}
