% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaProViz.R
\docType{package}
\name{MetaProViz}
\alias{MetaProViz-package}
\alias{MetaProViz}
\title{The MetaProViz package}
\description{
MetaProViz (Metabolomics Processing, functional analysis and
Visualization), a free open-source R-package that provides mechanistic
hypotheses from metabolomics data by integrating prior knowledge from
literature with metabolomics. MetaProViz offers an interactive
framework consisting of five modules: Processing, differential
analysis, prior knowledge access and refactoring, functional analysis
and visualization of both intracellular and exometabolomics
(consumption-release/core data).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://saezlab.github.io/MetaProViz}
  \item Report bugs at \url{https://github.com/saezlab/MetaProViz/issues}
}

}
\author{
Christina Schmidt \if{html}{\out{<\email{christina.schmidt@uni-heidelberg.de}>}}
and Jannik Franken and Denes Turei \if{html}{\out{<\email{turei.denes@gmail.com}>}}
and Dimitrios Prymidis and Macabe Daley and Julio Saez-Rodriguez
and Christian Frezza
}
