% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperLog.R
\name{MetaProViz_set_loglevel}
\alias{MetaProViz_set_loglevel}
\title{Sets the log level for the package logger}
\usage{
MetaProViz_set_loglevel(level, target = "logfile")
}
\arguments{
\item{level}{Character or class \code{loglevel}. The desired log level.}

\item{target}{Character, either 'logfile' or 'console'}
}
\value{
Returns \code{NULL}.
}
\description{
Sets the log level for the package logger
}
\examples{
MetaProViz_set_loglevel(logger::FATAL, target = 'console')

}
