% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPriorKnowledge.R
\name{metsigdb_chemicalclass}
\alias{metsigdb_chemicalclass}
\title{Metabolite chemical classes from RaMP DB}
\usage{
metsigdb_chemicalclass(version = "2.5.4", save_table = "csv", path = NULL)
}
\arguments{
\item{version}{\emph{Optional: } Version of the RaMP database loaded from OmniPathR. \strong{default: "2.5.4"}}

\item{save_table}{\emph{Optional: } File types for the analysis results are: "csv", "xlsx", "txt". \strong{Default = "csv"}}

\item{path}{{Optional:} String which is added to the resulting folder name \strong{default: NULL}}
}
\value{
A data frame containing the Prior Knowledge.
}
\description{
Metabolite chemical classes from RaMP DB
}
\examples{
ChemicalClass <- metsigdb_chemicalclass()

}
