% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPriorKnowledge.R
\name{metsigdb_kegg}
\alias{metsigdb_kegg}
\title{KEGG}
\usage{
metsigdb_kegg()
}
\value{
A data frame containing the KEGG pathways for ORA.
}
\description{
Import and process KEGG.
}
\details{
Imports KEGG pathways into the environment
}
\examples{
KEGG_Pathways <- MetaProViz::metsigdb_kegg()

}
