% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPriorKnowledge.R
\name{metsigdb_kegg}
\alias{metsigdb_kegg}
\title{KEGG pathways}
\usage{
metsigdb_kegg()
}
\value{
A data frame containing the KEGG pathways suitable for ORA.
}
\description{
KEGG pathways
}
\examples{
metsigdb_kegg()

}
