% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperSummarizedExperiment.R
\name{process_se}
\alias{process_se}
\title{Process SummarizedExperiment object to extract assay, sample, and feature metadata}
\usage{
process_se(se_obj)
}
\arguments{
\item{se_obj}{A SummarizedExperiment object.}
}
\value{
A list with elements:
\item{data}{Numeric matrix with metabolites as columns and samples as rows}
\item{metadata_sample}{Data frame with sample metadata (samples as row names)}
\item{metadata_feature}{Data frame with feature metadata (metabolites as row names)}
}
\description{
This helper function extracts the numeric data matrix, sample metadata, and feature metadata from a SummarizedExperiment object.
If the input is not a SummarizedExperiment, an error is thrown.
}
\examples{
Intra <- intracell_raw \%>\% tibble::column_to_rownames("Code")
mat <- Intra[-c(49:58), -c(1:5, 97)]
colData <- cellular_meta \%>\% dplyr::filter(Metabolite \%in\% colnames(mat)) \%>\% tibble::column_to_rownames("Metabolite")\%>\% .[colnames(mat), , drop = FALSE]
rowData <- Intra[-c(49:58), c(1:3)]
se <- SummarizedExperiment::SummarizedExperiment(assays = list(counts = mat), colData = colData, rowData = rowData)
result <- process_se(se)
str(result)

}
