% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing.R
\name{replicate_sum}
\alias{replicate_sum}
\title{Merges the analytical replicates of an experiment}
\usage{
replicate_sum(
  data,
  metadata_sample = NULL,
  metadata_info = c(Conditions = "Conditions", Biological_Replicates =
    "Biological_Replicates", Analytical_Replicates = "Analytical_Replicates"),
  save_table = "csv",
  path = NULL
)
}
\arguments{
\item{data}{SummarizedExperiment (se) file including assay and colData. If se file is provided, metadata_sample is extracted from the colData of the se object. Alternatively provide a DF which contains unique sample identifiers as row names and metabolite numerical values in columns with metabolite identifiers as column names. Use NA for metabolites that were not detected.}

\item{metadata_sample}{\emph{Optional: } Only required if you did not provide se file in parameter data. Provide DF which contains information about the samples, which will be combined with the input data based on the unique sample identifiers used as rownames. Must contain column with Conditions. If you do not have multiple conditions in your experiment assign all samples into the same condition. \strong{Default = NULL}}

\item{metadata_info}{\emph{Optional: } Named vector including the Conditions and Replicates information: c(Conditions="ColumnNameConditions", Biological_Replicates="ColumnName_metadata_sample", Analytical_Replicates="ColumnName_metadata_sample").\strong{Default = NULL}}

\item{save_table}{\emph{Optional: } File types for the analysis results are: "csv", "xlsx", "txt", ot NULL \strong{default: "csv"}}

\item{path}{\emph{Optional:} Path to the folder the results should be saved at. \strong{default: NULL}}
}
\value{
DF with the merged analytical replicates
}
\description{
Merges the analytical replicates of an experiment
}
\examples{
data(intracell_raw_se)
Res <- replicate_sum(data=intracell_raw_se)

data(intracell_raw)
Intra <- intracell_raw \%>\% tibble::column_to_rownames("Code")
Res <- replicate_sum(data=Intra[-c(49:58) ,-c(1:3)],
                                metadata_sample=Intra[-c(49:58) , c(1:3)])

}
\keyword{Analytical}
\keyword{Merge}
\keyword{Replicate}
