% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{ReplicateSum}
\alias{ReplicateSum}
\title{Merges the analytical replicates of an experiment}
\usage{
ReplicateSum(Input_data)
}
\arguments{
\item{Input}{Dataframe which contains unique sample identifiers as row names the Experimental design and the metabolite numerical values in columns with metabolite identifiers as column names. Needs to have Conditions, Biological_Replicates and Analytical_Replicate columns}
}
\description{
Merges the analytical replicates of an experiment
}
\keyword{Analyrical}
\keyword{Merge}
\keyword{Replicate}
