% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing.R
\name{ReplicateSum}
\alias{ReplicateSum}
\title{Merges the analytical replicates of an experiment}
\usage{
ReplicateSum(
  InputData,
  SettingsFile_Sample,
  SettingsInfo = c(Conditions = "Conditions", Biological_Replicates =
    "Biological_Replicates", Analytical_Replicates = "Analytical_Replicates"),
  SaveAs_Table = "csv",
  FolderPath = NULL
)
}
\arguments{
\item{InputData}{DF which contains unique sample identifiers as row names and metabolite numerical values in columns with metabolite identifiers as column names. Use NA for metabolites that were not detected.
#@param SettingsFile_Sample DF which contains information about the samples Column "Conditions", "Biological_replicates" and "Analytical_Replicates has to exist.}

\item{SettingsInfo}{\emph{Optional: } Named vector including the Conditions and Replicates information: c(Conditions="ColumnNameConditions", Biological_Replicates="ColumnName_SettingsFile_Sample", Analytical_Replicates="ColumnName_SettingsFile_Sample").\strong{Default = NULL}}

\item{SaveAs_Table}{\emph{Optional: } File types for the analysis results are: "csv", "xlsx", "txt", ot NULL \strong{default: "csv"}}

\item{FolderPath}{\emph{Optional:} Path to the folder the results should be saved at. \strong{default: NULL}}
}
\description{
Merges the analytical replicates of an experiment
}
\keyword{Analytical}
\keyword{Merge}
\keyword{Replicate}
