% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverRepresentationAnalysis.R
\name{StandardORA}
\alias{StandardORA}
\title{Uses enricher to run ORA on the differential metabolites (DM) using a pathway list}
\usage{
StandardORA(
  InputData,
  SettingsInfo = c(pvalColumn = "p.adj", PercentageColumn = "t.val", PathwayTerm =
    "term", PathwayFeature = "Metabolite"),
  pCutoff = 0.05,
  PercentageCutoff = 10,
  PathwayFile,
  PathwayName = "",
  minGSSize = 10,
  maxGSSize = 1000,
  SaveAs_Table = "csv",
  FolderPath = NULL
)
}
\arguments{
\item{InputData}{DF with metabolite names/metabolite IDs as row names. Metabolite names/IDs need to match the identifier type (e.g. HMDB IDs) in the PathwayFile.}

\item{SettingsInfo}{\emph{Optional: } Pass ColumnName of the column including parameters to use for pCutoff and PercentageCutoff. Also pass ColumnName for PathwayFile including term and feature names. (pvalColumn = ColumnName InputData, PercentageColumn= ColumnName InputData, PathwayTerm= ColumnName PathwayFile, PathwayFeature= ColumnName PathwayFile) \strong{c(pvalColumn="p.adj", PercentageColumn="t.val", PathwayTerm= "term", PathwayFeature= "Metabolite")}}

\item{pCutoff}{\emph{Optional: } p-adjusted value cutoff from ORA results. Must be a numeric value. \strong{default: 0.05}}

\item{PercentageCutoff}{\emph{Optional: } Percentage cutoff of metabolites that should be considered for ORA. Selects Top/Bottom \% of selected PercentageColumn, usually t.val or Log2FC \strong{default: 10}}

\item{PathwayFile}{DF that must include column "term" with the pathway name, column "Metabolite" with the Metabolite name or ID and column "Description" with pathway description that will be depicted on the plots.}

\item{PathwayName}{\emph{Optional: } Name of the PathwayFile used \strong{default: ""}}

\item{minGSSize}{\emph{Optional: } minimum group size in ORA \strong{default: 10}}

\item{maxGSSize}{\emph{Optional: } maximum group size in ORA \strong{default: 1000}}

\item{SaveAs_Table}{\emph{Optional: } File types for the analysis results are: "csv", "xlsx", "txt" \strong{default: "csv"}}

\item{FolderPath}{\emph{Optional:} Path to the folder the results should be saved at. \strong{default: NULL}}
}
\value{
Saves results as individual .csv files.
}
\description{
Uses enricher to run ORA on the differential metabolites (DM) using a pathway list
}
