% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyData.R
\docType{data}
\name{tissue_dma_young}
\alias{tissue_dma_young}
\title{tissue_dma_young}
\format{
Columns include Log2FC, stats, metabolite identifiers, metabolite pathways and normalised
metabolite values used as input with row names being metabolite trivial names.
}
\source{
Hakimi et. al, An integrated metabolic atlas of clear cell renal cell carcinoma, Cancer Cell 2016,
\doi{10.1016/j.ccell.2015.12.004}
}
\usage{
tissue_dma_young
}
\description{
We performed differential metabolite analysis comparing ccRCC tissue versus adjacent normal tissue of
the patient's subset of young patient's (age <42 years) using median normalised data from the supplementary
table 2 of Hakimi et. al.(="Tissue_Norm").
}
\keyword{datasets}
