% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyData.R
\docType{data}
\name{tissue_meta}
\alias{tissue_meta}
\title{Tissue_Metadata}
\format{
Columns include Metabolite IDs (HMDB, KEGG, etc), platform, mass, metabolic pathway with
row names being metabolite trivial names.
}
\source{
Hakimi et. al, An integrated metabolic atlas of clear cell renal cell carcinoma, Cancer Cell 2016,
\doi{10.1016/j.ccell.2015.12.004}
}
\usage{
tissue_meta
}
\description{
In Hakimi et. al. metabolites were assigned to metabolite IDs, pathways, platform, mass and other
fetaure metainformation.
}
\examples{
data(tissue_meta)
head(tissue_meta)
}
\keyword{datasets}
