% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyData.R
\docType{data}
\name{tissue_norm_se}
\alias{tissue_norm_se}
\title{tissue_norm_se}
\format{
Se object with numeric columns containing rmeasured metabolite (normalised data) intensities and
coldata include patient metadata (e.g. age, gender, sage, etc.)
}
\source{
Hakimi et. al, An integrated metabolic atlas of clear cell renal cell carcinoma, Cancer Cell 2016,
\doi{10.1016/j.ccell.2015.12.004}
}
\usage{
tissue_norm_se
}
\description{
This is median normalised data from the supplementary table 2 of Hakimi et al with metabolomic
profiling on 138 matched clear cell renal cell carcinoma (ccRCC)/normal tissue pairs.
}
\examples{
data(tissue_norm_se)
head(tissue_norm_se)
}
\keyword{datasets}
