% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyData.R
\docType{data}
\name{tissue_tvn_rnaseq}
\alias{tissue_tvn_rnaseq}
\title{tissue_tvn_rnaseq}
\format{
Columns include Log2FC, stats, gene name and SiRCle cluster information that summarises genes based
on their regulation
}
\source{
Mora & Schmidt, SiRCle (Signature Regulatory Clustering) model integration reveals mechanisms of phenotype
regulation in renal cancer, Genome Medicine 2024, \doi{10.1186/s13073-024-01415-3} Clark et. al, Integrated
proteogenomic characterization of clear cell renal cell carcinoma, Cell 2019, \doi{10.1016/j.cell.2019.10.007}
}
\usage{
tissue_tvn_rnaseq
}
\description{
The processed transcriptomics data was downloaded from the supplementary table 3 of Mora & Schmidt et. al., which
used the study from Clark et. al. under Proteomics data Commons PDC000127.
}
\examples{
data(tissue_tvn_rnaseq)
head(tissue_tvn_rnaseq)
}
\keyword{datasets}
