% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{ToyData}
\alias{ToyData}
\title{Access built-in example data}
\usage{
ToyData(name)
}
\arguments{
\item{name}{Character: name of a built-in dataset:
\itemize{
\item{\code{"IntraCells_Raw"}: }
\item{\code{"IntraCells_DMA"}: }
\item{\code{"CultureMedia_Raw"}: }
\item{\code{"Cells_MetaData"}: }
\item{\code{"Tissue_Norm"}: }
\item{\code{"Tissue_MetaData"}: }
}}
}
\value{
A data frame containing the toy data.
}
\description{
Import and process .csv file to create toy data.
}
\examples{
intra <- ToyData("IntraCells_Raw")

}
