% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyData.R
\name{ToyData}
\alias{ToyData}
\title{Access built-in example data}
\usage{
ToyData(Dataset)
}
\arguments{
\item{Dataset}{Character: name of a built-in dataset:
\itemize{
\item{\code{"IntraCells_Raw"}: }
\item{\code{"IntraCells_DMA"}: }
\item{\code{"CultureMedia_Raw"}: }
\item{\code{"Cells_MetaData"}: }
\item{\code{"Tissue_Norm"}: }
\item{\code{"Tissue_MetaData"}: }
\item{\code{"Tissue_DMA"}: }
\item{\code{"Tissue_DMA_Old"}: }
\item{\code{"Tissue_DMA_Young"}: }
\item{\code{"Tissue_TvN_Proteomics"}: }
\item{\code{"Tissue_TvN_RNAseq"}: }
\item{\code{"EquivalentFeatures"}: }
}}
}
\value{
A data frame containing the toy data.
}
\description{
Import and process .csv file to create toy data DF.
}
\examples{
ToyData("IntraCells_Raw")

}
