% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefactorPriorKnoweldge.R
\name{TranslateID}
\alias{TranslateID}
\title{Translate IDs to/from KEGG, PubChem, Chebi, HMDB}
\usage{
TranslateID(
  InputData,
  SettingsInfo = c(InputID = "MetaboliteID", GroupingVariable = "term"),
  From = c("kegg"),
  To = c("pubchem", "chebi", "hmdb"),
  SaveAs_Table = "csv",
  FolderPath = NULL
)
}
\arguments{
\item{InputData}{Dataframe with at least one column with the target (e.g. metabolite), you can add other columns such as source (e.g. term)}

\item{SettingsInfo}{\emph{Optional: } Column name of Target in Input_GeneSet. \strong{Default = list(InputID="MetaboliteID" , GroupingVariable="term")}}

\item{SaveAs_Table}{\emph{Optional: } File types for the analysis results are: "csv", "xlsx", "txt". \strong{Default = "csv"}}

\item{FolderPath}{{Optional:} String which is added to the resulting folder name \strong{Default = NULL}}
}
\value{
List with three DFs: 1) Original data and the new column of translated ids. 2) Mapping summary from Original ID to Translated. 3) Mapping summary from Translated to Original.
}
\description{
Translate IDs to/from KEGG, PubChem, Chebi, HMDB
}
\examples{
KEGG_Pathways <- MetaProViz::LoadKEGG()
res <- MetaProViz::TranslateID(InputData= KEGG_Pathways, SettingsInfo = c(InputID="MetaboliteID", GroupingVariable="term"), From = c("kegg"), To = c("pubchem","chebi","hmdb"), Method="GetAll", SaveAs_Table= "csv", FolderPath=NULL)

}
\keyword{IDs}
\keyword{Translate}
