% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizHeatmap.R
\name{VizHeatmap}
\alias{VizHeatmap}
\title{Heatmap visualization}
\usage{
VizHeatmap(
  data,
  metadata_info = NULL,
  metadata_sample = NULL,
  metadata_feature = NULL,
  plot_name = "",
  scale = "row",
  save_plot = "svg",
  enforce_featurenames = FALSE,
  enforce_samplenames = FALSE,
  print_plot = TRUE,
  path = NULL
)
}
\arguments{
\item{data}{DF with unique sample identifiers as row names and
metabolite numerical values in columns with metabolite identifiers as column
names. Includes experimental design and outlier column.}

\item{metadata_info}{\emph{Optional: } NULL or Named vector  where you can
include vectors or lists for annotation c(individual_Metab=
"ColumnName_metadata_feature",individual_Sample=
"ColumnName_metadata_sample",
color_Metab="ColumnName_metadata_feature", color_Sample=
list("ColumnName_metadata_sample",
"ColumnName_metadata_sample",...)).\strong{Default = NULL}}

\item{metadata_sample}{DF which contains information about the samples,
which will be combined with your input data based on the unique sample
identifiers. and other columns with required plot_typeInfo.\strong{Default
= NULL}}

\item{metadata_feature}{\emph{Optional: } DF with column "Metabolite"
including the Metabolite names (needs to match Metabolite names of
Input_data) and other columns with required plot_typeInfo. \strong{Default
= NULL}}

\item{plot_name}{\emph{Optional: } String which is added to the output files of the plot}

\item{scale}{\emph{Optional: } String with the information for scale row, column or none. \strong{Default = row}}

\item{save_plot}{\emph{Optional: } Select the file type of output plots. Options are svg, pdf, png or NULL. \strong{Default = "svg"}}

\item{enforce_featurenames}{\emph{Optional: } If there are more than 100 features no rownames will be shown, which is due to readability. You can Enforce this by setting this parameter to TRUE. \strong{Default = FALSE}}

\item{enforce_samplenames}{\emph{Optional: } If there are more than 50 sampless no colnames will be shown, which is due to readability. You can Enforce this by setting this parameter to TRUE. \strong{Default = FALSE}}

\item{print_plot}{\emph{Optional: } print the plots to the active graphic
device.}

\item{path}{{Optional:} String which is added to the resulting folder name \strong{default: NULL}}
}
\value{
List with two elements: Plot and Plot_Sized
}
\description{
Heatmap visualization
}
\examples{
Intra <- ToyData("IntraCells_Raw")
Res <- MetaProViz::VizHeatmap(data=Intra[,-c(1:3)])

}
\keyword{Heatmap}
