% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizHeatmap.R
\name{VizHeatmap}
\alias{VizHeatmap}
\title{This script allows you to perform different data visualizations using the results of the MetaProViz analysis}
\usage{
VizHeatmap(
  Input_data,
  Plot_SettingsInfo = NULL,
  Plot_SettingsFile_Sample = NULL,
  Plot_SettingsFile_Metab = NULL,
  OutputPlotName = "",
  SCALE = "row",
  Save_as_Plot = "svg",
  enforce_FeatureNames = FALSE,
  enforce_SampleNames = FALSE
)
}
\arguments{
\item{Input_data}{DF with unique sample identifiers as row names and metabolite numerical values in columns with metabolite identifiers as column names. Includes experimental design and outlier column.}

\item{Plot_SettingsInfo}{\emph{Optional: } NULL or Named vector  where you can include vectors or lists for annotation c(individual_Metab= "ColumnName_Plot_SettingsFile_Metab",individual_Sample= "ColumnName_Plot_SettingsFile_Sample", color_Metab="ColumnName_Plot_SettingsFile_Metab", color_Sample= list("ColumnName_Plot_SettingsFile_Sample", "ColumnName_Plot_SettingsFile_Sample",...)).\strong{Default = NULL}}

\item{Plot_SettingsFile_Sample}{DF which contains information about the samples, which will be combined with your input data based on the unique sample identifiers. and other columns with required PlotSettingInfo.\strong{Default = NULL}}

\item{Plot_SettingsFile_Metab}{\emph{Optional: } DF with column "Metabolite" including the Metabolite names (needs to match Metabolite names of Input_data) and other columns with required PlotSettingInfo. \strong{Default = NULL}}

\item{OutputPlotName}{\emph{Optional: } String which is added to the output files of the plot}

\item{SCALE}{\emph{Optional: } String with the information for scale row or column. \strong{Default = row}}

\item{Save_as_Plot}{\emph{Optional: } Select the file type of output plots. Options are svg, pdf or png. \strong{Default = "svg"}}

\item{enforce_FeatureNames}{\emph{Optional: } If there are more than 100 features no rownames will be shown, which is due to readability. You can enforce this by setting this parameter to TRUE. \strong{Default = FALSE}}

\item{enforce_SampleNames}{\emph{Optional: } If there are more than 50 sampless no colnames will be shown, which is due to readability. You can enforce this by setting this parameter to TRUE. \strong{Default = FALSE}}
}
\description{
This script allows you to perform different data visualizations using the results of the MetaProViz analysis
}
\keyword{Volcano}
\keyword{pathways}
\keyword{plot,}
