% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{VizPCA}
\alias{VizPCA}
\title{This script allows you to perform different data visualizations using the results of the MetaProViz analysis}
\usage{
VizPCA(
  Input_data,
  Color = FALSE,
  Shape = FALSE,
  Show_Loadings = FALSE,
  Scaling = TRUE,
  Theme = theme_classic(),
  OutputPlotName = "",
  Save_as = "svg"
)
}
\arguments{
\item{Input_data}{DF with unique sample identifiers as row names and metabolite numerical values in columns with metabolite identifiers as column names. Use NA for metabolites that were not detected. includes experimental design and outlier column.}

\item{Color}{\emph{Optional: }String which contains the name of the output file of the Metabolic Clusters}

\item{Shape}{\emph{Optional: }String which contains the name of the output file of the Metabolic Clusters}

\item{Show_Loadings}{\emph{Optional: } TRUE or FALSE for whether PCA loadings are also plotted on the PCA (biplot) \strong{Default = FALSE}}

\item{Scaling}{\emph{Optional: } TRUE or FALSE for whether a data scaling is used \strong{Default = TRUE}}

\item{Theme}{\emph{Optional: } Selection of theme for plots from ggplot2. \strong{Default = theme_classic} ??}

\item{OutputPlotName}{\emph{Optional: } String which is added to the output files of the PCA}

\item{Save_as}{\emph{Optional: } Select the file type of output plots. Options are svg, png, pdf, jpeg, tiff, bmp. \strong{Default = svg}}
}
\description{
This script allows you to perform different data visualizations using the results of the MetaProViz analysis
}
\keyword{PCA}
