% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{VizPCA}
\alias{VizPCA}
\title{This script allows you to perform different data visualizations using the results of the MetaProViz analysis}
\usage{
VizPCA(
  Plot_SettingsInfo = NULL,
  Plot_SettingsFile = NULL,
  Input_data,
  color_palette = NULL,
  shape_palette = NULL,
  Show_Loadings = FALSE,
  Scaling = TRUE,
  Theme = NULL,
  OutputPlotName = "",
  Save_as_Plot = "svg",
  Save_as_Results = "csv"
)
}
\arguments{
\item{Plot_SettingsInfo}{\emph{Optional: } NULL or Named vector including at least one of those three information : c(color="ColumnName_Plot_SettingsFile", shape= "ColumnName_Plot_SettingsFile"). \strong{Default = NULL}}

\item{Plot_SettingsFile}{\emph{Optional: } DF which contains information about the samples, which will be combined with your input data based on the unique sample identifiers used as rownames. Column "Conditions" with information about the sample conditions (e.g. "N" and "T" or "Normal" and "Tumor"), can be used for feature filtering and colour coding in the PCA. Column "AnalyticalReplicate" including numerical values, defines technical repetitions of measurements, which will be summarised. Column "BiologicalReplicates" including numerical values. Please use the following names: "Conditions", "Biological_Replicates", "Analytical_Replicates".\strong{Default = NULL}}

\item{Input_data}{DF with a column "UniqueID" with unique sample identifiers  and metabolite numerical values in columns with metabolite identifiers as column names. Use NA for metabolites that were not detected. includes experimental design and outlier column.}

\item{color_palette}{\emph{Optional: } Provide customiced color-palette in vector format. \strong{Default = NULL}}

\item{shape_palette}{\emph{Optional: } Provide customiced shape-palette in vector format. \strong{Default = NULL}}

\item{Show_Loadings}{\emph{Optional: } TRUE or FALSE for whether PCA loadings are also plotted on the PCA (biplot) \strong{Default = FALSE}}

\item{Scaling}{\emph{Optional: } TRUE or FALSE for whether a data scaling is used \strong{Default = TRUE}}

\item{Theme}{\emph{Optional: } Selection of theme for plots from ggplot2. \strong{Default = NULL}}

\item{OutputPlotName}{\emph{Optional: } String which is added to the output files of the PCA \strong{Default = ""}}

\item{Save_as_Plot}{\emph{Optional: } Select the file type of output plots. Options are svg, png, pdf. \strong{Default = svg}}

\item{Save_as_Results}{\emph{Optional: } File types for the analysis results are: "csv", "xlsx", "txt" \strong{default: "csv"}}
}
\description{
This script allows you to perform different data visualizations using the results of the MetaProViz analysis
}
\keyword{PCA}
