% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizSuperplots.R
\name{VizSuperplot}
\alias{VizSuperplot}
\title{This script allows you to perform different visualizations (bar, box, violin plots) using the results of the MetaProViz analysis}
\usage{
VizSuperplot(
  Input_data,
  Experimental_design,
  Graprh_Style = "Box",
  Superplot = NULL,
  OutputPlotName = "",
  Output_plots = "Together",
  Selected_Conditions = NULL,
  Selected_Comparisons = NULL,
  Theme = theme_classic(),
  Save_as_Plot = "svg"
)
}
\arguments{
\item{Input_data}{DF with unique sample identifiers as row names and metabolite numerical values in columns with metabolite identifiers as column names. Includes experimental design and outlier column.}

\item{Experimental_design}{DF which contains information about the samples, which will be combined with your input data based on the unique sample identifiers used as rownames. Column "Conditions" with information about the sample conditions (e.g. "N" and "T" or "Normal" and "Tumor"), can be used for feature filtering and colour coding in the PCA. Column "AnalyticalReplicate" including numerical values, defines technical repetitions of measurements, which will be summarised. Column "BiologicalReplicates" including numerical values. Please use the following names: "Conditions", "Biological_Replicates", "Analytical_Replicates".}

\item{Graprh_Style}{String with the information of the Graph style. Available options are Bar. Box and Violin  \strong{Default = Box}}

\item{Superplot}{\emph{Optional: } String with a Column name of the Experimental_design as string which is used to make the plots Superplots.}

\item{Output_plots}{String with plot save information. Available options are "Individual" for plots of each Individual metabolite and "Together" for a pdf containing all the plots. \strong{Default = Together}}

\item{Selected_Conditions}{Vector with names of selected Conditions for the plot. \strong{Default = NULL}}

\item{Selected_Comparisons}{Logical, TRUE to use t.test between the Selected_Conditions or FALSE. \strong{Default = NULL}}

\item{Theme}{\emph{Optional: } Selection of theme for plot. \strong{Default = theme_classic} ??}

\item{Save_as_Plot}{\emph{Optional: } Select the file type of output plots. Options are svg, pdf or png. \strong{Default = svg}}

\item{Output_Name}{\emph{Optional: } String which is added to the output files of the plot.}
}
\description{
This script allows you to perform different visualizations (bar, box, violin plots) using the results of the MetaProViz analysis
}
\keyword{Barplot,}
\keyword{Boxplot,}
\keyword{Superplot}
\keyword{Violinplot,}
